Moving sync builders from the FYI waterfall to the main chromium waterfall
The sync builders "Chromium Windows7 Sync", "Chromium Mac Sync" and "Chromium Linux Sync" are builders that run the sync integration tests on each chromium checkin, and have lived on the FYI waterfall for the past few quarters. Now that the integration tests are mature and stable and the tests have been green for a while, it is time to move the builders over to the main chromium waterfall.

This patch reconfigures the master and slave cfg files in both, master.chromium and master.chromium.fyi.

BUG=70060
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6325007

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@72036 0039d316-1c4b-4281-b951-d872f2087c98