V8 buildbot update, remove crankshaft, add isolates and nosse*
In addition to adding a slave running the isolates test this includes
tree new slaves runing without sse2, sse3, and sse4.1. V8testing and
the factory files has been clean up for the simulator flag which is
not used anyway (except for when we compile, but it is already
provided there). 
Review URL: http://codereview.chromium.org/6770008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@79795 0039d316-1c4b-4281-b951-d872f2087c98