A few changes required for the NAS support.
Add a 'parentname' build property on the testers containing the
name of the builder that triggered them. Make sure we send it with
the build_properties to the slave commands.

Always propagate the "snapshot" property during Trigger if the
builder is a "NASBuilder".

Add the definition for the AddCloneStep and AddSnapshotStep steps.

Update gclient factory to add the snapshot and clone steps based
on the slave type: NASTester or NASBuilder.
Review URL: http://codereview.chromium.org/6823029

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@81265 0039d316-1c4b-4281-b951-d872f2087c98