Modify the way we pass the --results-directory to run_webkit_tests
In order to accomodate the fix in https://bugs.webkit.org/show_bug.cgi?id=58272,
we will temporarily pass --results-directory as an absolute path.

This fix needs to land *after* that fix. We will then land a subsequent fix
to webkit.org changing the way relative paths are interpreted for chromium
ports, and then remove the hack in this code.
Review URL: http://codereview.chromium.org/6834004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@81662 0039d316-1c4b-4281-b951-d872f2087c98