Purge Buildbot of its BuildStep checks.
Starting with Buildbot 0.8 there are now checks on
addStep() which verify a given step is related to a
BuildStep.  These checks trip the log parsers up since
we are actually passing a currying instance through
addStep that should just be called by Buildbot.

With the checks in place, Buildbot fails to
start when any step is added that's based on
InitializePartiallyWithArguments.  All of our
performance tests, gtests, and layout tests are based
on this class.

I experimented with tricking out inheritance to avoid
needing to change Buildbot code at all.  I don't
want to waste any more time on those types of 'fixes'.
IPWA should be removed, and it should be removed soon,
not be given a lifeline.

However, I'm open to real suggestions if anyone has
them.  I'd prefer not to have to modify this Buildbot
code, so offer those suggestions up if you have them.
For this change not to be needed, this condition must
be true:

  not isinstance(step_or_factory, BuildStep) and
  type(step_or_factory) == type(BuildStep) and
  issubclass(step_or_factory, BuildStep)

where type(BuildStep) == 'classobj'.
type(step_or_factory) is currently 'instance'.

The right fix: get rid of the complicated log parsers.
We're getting there, though it takes a lot more work
which I'm unavailable to do right now.  When that's
done, we'll be able to treat all of the log parsing
steps as regular BuildSteps and say goodbye to IPWA.

For now, we'll just keep IPWA in the basement and make
sure to keep the door locked.

BUG=79195

Review URL: http://codereview.chromium.org/7206017

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@89690 0039d316-1c4b-4281-b951-d872f2087c98