Archive clobber builds in Google Storage.
This change modifies the Chromium master config while
modifying archive_build to disregard this change once
the master is restarted with it.  When we are ready
to test this change on the master, we can remove just
those lines from archive_build.py to verify it.

Patch from Stefan Zager <szager@chromium.org>.
CL originally reviewed at
http://codereview.chromium.org/7190039/.

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@90114 0039d316-1c4b-4281-b951-d872f2087c98