Enable auto-reboot for all perf bots.
Many of these bots are now in a state where
they frequently require reboots because the test
infrastructure and/or Chrome itself are causing
the system strain when Chrome is run repeatedly.
Rebooting the system will clear local symptoms of this
problem and allow the tests to complete in a timely
manner without leading to unsightly hangs or timeouts.

Review URL: http://codereview.chromium.org/7082008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@90689 0039d316-1c4b-4281-b951-d872f2087c98