Reorder the linux slaves.
This has an incidence on which slave runs which builder. By using a rolling
setup instead of slices, it makes it easier to add more slaves. Otherwise,
adding a new slave will change configuration of the previous slaves. This won't
be the case anymore after this change.

R=nsylvain@chromium.org
BUG=
TEST=


Review URL: http://codereview.chromium.org/7461047

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@93806 0039d316-1c4b-4281-b951-d872f2087c98