V8 buildbot: Refactor the v8 master to make it easier to change the currently tracked branches.
The current code has the branches that we are currently tracking
spread out all over the code in the master.cfg and slaves.cfg
files. This means that we need to manually update a lot of places
every 6 weeks, plus cleanup on the slaves to delete the old
directories.

The new approach puts the branch names into a seperate file so that it
can be loaded from both master.cfg and slaves.cfg.
Review URL: http://codereview.chromium.org/7566034

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@95582 0039d316-1c4b-4281-b951-d872f2087c98