Add Skia webkit buildbots alongside all CG webkit buildbots.
For now, we want to run BOTH Core Graphics and Skia; once we see Skia working
for webkit tests and Chromium as a whole, we will turn off the Core Graphics
buildbots.

The same buildbot slaves are used by both CG and Skia builds; as long as we
are running both types of builds, they will alternate back and forth.
Review URL: http://codereview.chromium.org/7749012

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@98634 0039d316-1c4b-4281-b951-d872f2087c98