Add the ability (optional) to patch the chromite tools for a given queue.
This allows changes to the chromite tools to be tested in a realistic setting,
but under the careful control of the master config.  The format used should
allow tools that fail back to being up-to-date rather than patched and
stagnant, without triggering the branch-checking safety checks.

BUG=chromium-os:20164
TEST=Run Locally.

Review URL: http://codereview.chromium.org/7789016

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@100065 0039d316-1c4b-4281-b951-d872f2087c98