Promote the ASAN bots to be tree closer.
Since the main waterfall is too crowded, we need a separate
memory waterfall for the ASAN bot. One that will actually close the
tree if something goes red.

Since we already had a memory waterfall, I renamed it memory.fyi, and
created another memory waterfall.

Eventually, once a bot of the memory.fyi waterfall is stable, green and
cycle within ~30 minutes, I'd like that bot to join the ASAN bots on the
memory waterfall, slowly reducing the number of bots on the MemoryFyi
waterfall.

The main tree sheriff should be in change of the memory waterfall, while
the current memory sheriffs will take care of the memory.fyi waterfall.

The name of the directory for the new memory waterfall is
master.chromium.memory.new because svn did not let me move a directory and
create a new one with a similar name in the same change. We'll need to
make a follow up patch to rename master.chromium.memory.new to
master.chromium.memory.

This is a port of http://codereview.chromium.org/7967019/.

TBR=nsylvain@chromium.org

Review URL: http://codereview.chromium.org/8043019

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@102695 0039d316-1c4b-4281-b951-d872f2087c98