Add new asan_archive_build.py file.
This will eventually be used to archive an asan package and upload it to
google storage where the security team will be able to test with it.

For now, this is only creating an archive. The code is inspired from
zip_build.py, but I suspect the common code will eventually go away
as we make more progress.

Review URL: http://codereview.chromium.org/8362014

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@106625 0039d316-1c4b-4281-b951-d872f2087c98