Don't close the chromiumOS tree for failures in BuildTarget.
This is compiled from the LKGM, hence we have built with a previous version
of chrome successfully on all of these platforms.  Stopping changes in
chromiumOS wont make any problems easier to fix.  The Chrome Sheriffs should
pass the buck along if the end result is that chromiumOS is historically
broken.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/8404021

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@107521 0039d316-1c4b-4281-b951-d872f2087c98