Set a default subject to maintain forward compatibility.
Buildbot 0.7.12 sets a default subject in MailNotifier
like in this CL.  Buildbot 0.8.4p1 sets a default
subject in MailNotifier that uses a key named title.
If a Chromium master was set up on 0.7.12, used
ChromiumNotifier, and that master did not set a subject,
it would get the 0.7.12 MailNotifier default that used
all the keys that ChromiumNotifier sets, including
projectName (but not title).

If that master was later converted to Buildbot 0.8.4,
it would get a new default subject but would not get
an updated set of keys to allow converting the subject
into a proper string.

Fix this by setting the default subject to match the
0.7.12 default in MailNotifier.  I considered setting
the title key, this seems less error prone.


Review URL: http://codereview.chromium.org/8935010

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@114258 0039d316-1c4b-4281-b951-d872f2087c98