Add a marker to explicitly specify that the default tests must be run.
This way, it could be specified in PRESUBMIT.py to be clear that default tests
should be run. This marker is implicit if bot=builder_name is specified alone.
Use a DFA to maintain the string parsing context.

Use set() to filter out duplicate tests.

R=petermayo@chromium.org
BUG=
TEST=


Review URL: http://codereview.chromium.org/9431045

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@123241 0039d316-1c4b-4281-b951-d872f2087c98