Add back proper clobber support
Check for clobber was lost in the refactoring. It's still working because when
clobber=foo is provided, the string 'foo' is considered True. That's because
trychange.py never provides clobber=false. With this change, TryJobBase will try
to intrepret the string instead of just looking for its presence.

R=petermayo@chromium.org
BUG=
TEST=


Review URL: http://codereview.chromium.org/9447027

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@123302 0039d316-1c4b-4281-b951-d872f2087c98