Make step names not include squared comments.
The chromiumos trees want to repeat steps with a board designation
tacked on to instances to distinguish them, but the messaging and
notification of these messages is not sensitive to which board. So with
this change the configuration will be insensitive to the extra stuff
in the step name.

R=cmp@chromium.org,maruel@chromium.org
BUG=115201
TEST=unittests

Review URL: http://codereview.chromium.org/9701022

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@126699 0039d316-1c4b-4281-b951-d872f2087c98