Add support for setting the ACL when uploading to Google Storage.
This CL modifies the helper methods for uploading files to Google Storage so
they accept a canned-acl value (see gsutil --help for canned-acl values) when
uploading a file.

slave_utils.GSUtilCopyFile is where the initial change is made, and the rest
of the files in this CL are the uses of that method. (With the exception of
gsutil_cp_dir.py, which calls gsutil directly, but is not used anywhere else.)

The implementation changes the default behavior so that if no canned-acl is
specified then no -a option is sent to gsutil, resulting in the uploaded file
getting the bucket's default object ACL or 'project-private' if no default
object ACL is set.

I've made the slave scripts default to 'public-read' since that was the
hard-coded behavior before this change, so things will continue working as-is
without master restarts. I've also included the necessary changes in the
master.cfg files, so that after the masters are restarted, I can go back and
change the default value in the slave script from 'public-read' to None.

BUG=110488
TEST=None

Review URL: http://codereview.chromium.org/9253015

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@128527 0039d316-1c4b-4281-b951-d872f2087c98