Add no-new-builds command to stop starting new builds.
After this CL, masters started will support handling the
'make no-new-builds' command.  This command will cause
the master to move into a degraded state where new builds
will not be started.  All schedulers are removed, all
sources are removed, and all pending builds are canceled.

The normal way to recover from these changes will be to
stop and restart the master.  If a restart was planned
and no-new-builds is run, the restart will need to happen
even if the previous need for it went away.

R=maruel@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9950049

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@131042 0039d316-1c4b-4281-b951-d872f2087c98