Use the sharding supervisor on the main waterfall.
Modify the sharding supervisor to take a whitelist of tests to shard instead
of a blacklist, this make it easier to do a controlled deployment of the
sharding supervisor.

Also make it understand when slaves only run a subset of a test suite.

This should be a no_op for the try server.

Review URL: http://codereview.chromium.org/9979017

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@131274 0039d316-1c4b-4281-b951-d872f2087c98