Delete snapshots more than a day old.
This goes hand-in-hand with https://chromiumcodereview.appspot.com/10007043/, which will generate snapshots for Chrome Frame tests that time out.  This gives diagnosticians a day to retrieve snapshots before they're reaped.  In my dream world, the snapshots are uploaded somewhere and automatically linked to in the log for the test that timed out.  The thought makes me giddy with excitement.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/9972002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@132593 0039d316-1c4b-4281-b951-d872f2087c98