Stop hotpatching the slave when it is run by a developper.
Use 'TESTING_MASTER' not in os.environ to detect when the slave is run as a
normal slave. In the contrary, do not hot patch the slave to automatically
delete unknown directories in the current directory or to automatically reboot
the slave when done.

R=petermayo@chromium.org
BUG=
TEST=


Review URL: http://codereview.chromium.org/10363005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@135270 0039d316-1c4b-4281-b951-d872f2087c98