Fix getLastFinishedBuild() so it works in multi-slave per builder cases.
Fix getLastFinishedBuild() so it works correctly when there are 2 or more
running builds for a given builder.  Previously, it would always return the
second-to-last build if the first was missing or running.  Now it always
returns the build that is just after what we assert is the oldest running
build.

BUG=126251
R=maruel@chromium.org,szager@chromium.org


Review URL: http://codereview.chromium.org/10365024

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@135489 0039d316-1c4b-4281-b951-d872f2087c98