Fix LKGR algorithm.
My changes in r135570 and r135574 were syntactically correct
but semantically incorrect.  The problem with them was that the
LKGR algorithm uses continue/break to maneuver through its
nested for loops, and r135570 added a layer of nesting which
invalidated the layers of nesting.

This change restores the original flow control through the
loops, and makes the LKGR algorithm work as it did before
r135570 except now multiple masters should be supported.

TBR=szager@chromium.org


Review URL: https://chromiumcodereview.appspot.com/10310057

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@135819 0039d316-1c4b-4281-b951-d872f2087c98