Add a master.chromiumos.tryserver.
This will temporarily be a debug master on master2, which polls
from the 'test' branch of the tryjobs repo.  Once we've tested this,
we'll switch most of the trybots over from master4 to master2, and
then make this one the production host.

Corresponding build_internal change is at
https://chromereviews.googleplex.com/4531016/

BUG=chromium-os:30344
TEST=Ran master locally

Review URL: https://chromiumcodereview.appspot.com/10517013

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@140621 0039d316-1c4b-4281-b951-d872f2087c98