Add vm47-1 as an Android Latest bot on the WebKit Canary master
Introduce ChromiumWebkitLatestAnnotationFactory which combines the needs of the
Android builders with the characteristics of Webkit Latest builders. A new
category will be appended at the end of the WebKit Canary master containing
only this builder, mimicing the Android builder on the main Chromium waterfall.

BUG=132306
TEST=

Review URL: https://chromiumcodereview.appspot.com/10539116

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@142123 0039d316-1c4b-4281-b951-d872f2087c98