Add a builder to compile Chromium iOS for devices.
The build was only running simulator builds, but we want to run both simulator and device because there are subtle differeces.

The builder definitions have a significant amount of duplication, but they will diverge as we get tests running so I'd like to keep them separate from the get go.

Review URL: https://chromiumcodereview.appspot.com/10802084

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@148789 0039d316-1c4b-4281-b951-d872f2087c98