Update runtest.py to run iOS test apps in the simulator.
The script uses iossim to launch the simulator and run the test app. After the app completes, the script prints any new crash logs that may exist (i.e. in case the app crashed).

This is a re-apply of https://chromiumcodereview.appspot.com/10807011/ but it should no longer break mac builds. The problem was that the code was checking
  if test_platform in ('ios-simulator'):
but since there was no trailing comma, it was treating it as a string instead of a tuple. This resulted in mac builds calling main_ios because they don't specify a platform (and '' in 'any string' == True).

Review URL: https://chromiumcodereview.appspot.com/10821115

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@149629 0039d316-1c4b-4281-b951-d872f2087c98