Fix extract_build to also replace variables in passed URLs.
Originally, extract_build would replace variables in URLs
so that when --build-url was used and contained a
variable, this would also be updated to include the value
of that variable.  During a recent refactor, the code
changed in such a way that this no longer worked.

Update the code to add support for that back to extract_build.

TBR=kerz@chromium.org


Review URL: https://chromiumcodereview.appspot.com/10831197

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@150393 0039d316-1c4b-4281-b951-d872f2087c98