Update run_slavelastic to Handle Network Disconnects.
Fixing an issue where the slave was unable to open the new zip file (I assume the network path was down for a second, but not sure) by catching the error and nicely aborting. This gives the tests that still need to be processed after a chance.

BUG=


Review URL: https://chromiumcodereview.appspot.com/10832341

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@152590 0039d316-1c4b-4281-b951-d872f2087c98