Add script to download Android images from google storage and flash devices.
The script is a stand-alone script to download zip images from google storage
using gsutil.  It requires gsutil to be installed and the slave machint to
appropriately register credentials for access to the images.  The script has
logic to only flash devices that need to be flashed and can be flashed with
the specified build.  It has been tested locally on my machine after I
downloaded and installed gsutil and authenticated my credentials.

BUG=142811

Review URL: https://chromiumcodereview.appspot.com/10893011

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@157420 0039d316-1c4b-4281-b951-d872f2087c98