Create option for a BuildStep to be suppressed from waterfall.
This is a cherry-pick of a feature landed in BuildBot 0.8.6, which allows for a
hideStepIf mechanism that hides steps from the waterfall. This is needed for
buildrunner integration, in that steps are skipped and hidden from the real
waterfall, but enabled when run from the buildrunner/annotator.

The original commits are here:
https://github.com/buildbot/buildbot/commit/2577dac35b39c4ee9c419681021d7971e8f1e5d2
https://github.com/buildbot/buildbot/commit/3db80d46d3c35d71591f033e5c3d0a2fa0001c59

BUG=


Review URL: https://chromiumcodereview.appspot.com/10961025

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@159833 0039d316-1c4b-4281-b951-d872f2087c98