[sync] Shard sync_integration_tests on Linux trybots and buildbot
The sync_integration_tests target takes approximately 20 minutes to run on the chromium buildbots and trybots. This is because each of the 250+ tests spins up a python test server and brings up a full-fledged instance of chrome with all services running.

This patch enables sharding for the tests on Linux. Local tests seem to indicate that we can get a 70% reduction in run time (~20 minutes down to ~6 minutes).

BUG=157234
TEST=sync_integration_tests

Review URL: https://chromiumcodereview.appspot.com/11231057

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@163419 0039d316-1c4b-4281-b951-d872f2087c98