Add step seeding and step cursor support to annotator.
This allows us to have more flexible buildsystems on the slave, including ones
that operate steps in parallel (swarm). Step seeding allows us to send a
manifest of steps we are about to execute to pre-populate the buildbot
waterfall. A further extension would allow seeded steps to be marked
tree-closing, providing an annotated solution to gatekeeper. Step cursor allows
us to specify which step the subsequent annotator commands apply to, which allow
us to take advantage of step seeding and also enable parallel execution.
Previous code that relies on BUILD_STEP should be unaffected.

This is a resubmission of https://chromiumcodereview.appspot.com/11109007/,
which caused issues with existing buildscripts. This version uses STEP_CLOSED
to explicitly close an annotator section, allowing for backwards compatibility.

BUG=152101


Review URL: https://chromiumcodereview.appspot.com/11229052

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@164786 0039d316-1c4b-4281-b951-d872f2087c98