Move Chrome Endure slaves to a new master called ChromiumEndure.
The Chrome Endure slaves previously ran on the ChromiumPyauto master.
That master is going to be trimmed down, and part of doing so is to
move the Chrome Endure bots off that master.  The Chrome Endure bots
will now be run on a new master called ChromiumEndure.

BUG=chromium-os:35235
TEST=Verified the ChromiumPyauto and ChromiumEndure masters can start up
     successfully when run on my local machine.

Review URL: https://chromiumcodereview.appspot.com/11094085

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@166032 0039d316-1c4b-4281-b951-d872f2087c98