Selective build clobbering feature (compile.py).
Adds the ability for devs/troopers/etc. to set 'landmines' in the tree so that
the build will selectively clobber when a builder moves over a revision with such
a change.

This cl hooks the clobber mechanism to the compile.py build script.

The related cl for android's build scripts and an initial landmines.py is here:
https://chromiumcodereview.appspot.com/11175016/

R=cmp,maruel@chromium.org
BUG=121897

Review URL: https://chromiumcodereview.appspot.com/11234013

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@167551 0039d316-1c4b-4281-b951-d872f2087c98