Make runtest.py's --run-python-script argument actually work.
The runtest script checks that the given test_exe lives in the
build directory. This broke --run-python-script because python scripts
do not live in the build dir.

There was only one user of --run-python-script, which only used it on
Windows, which worked because the build-dir and target was given as ''.
For the other platforms, the _GetPythonTestCommand function worked around
the problem. This function has been removed.
Review URL: https://codereview.chromium.org/11293200

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@167931 0039d316-1c4b-4281-b951-d872f2087c98