Chromium WebRTC ToT waterfall
A waterfall that builds Chromium with WebRTC ToT on Windows, Mac and Linux. It has separate builders and testers but I guess these can be configured on the same machines to begin with.

One caveat is that the Tester machines must have a webcam available. It's fine using a virtual webcam by installing software, but then this needs to be installed manually since that's an addition to the current OS images.

BUG=None
TEST=Successful builds with local master and slaves with Linux, Mac and Windows (except for the 'packaged build' step for Windows, since I lack gsutil credentials for the bucket). Testers fail on 'extract_build' but I assume that just requires proper setup of hostname/share directories?

Review URL: https://codereview.chromium.org/11175053

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@167943 0039d316-1c4b-4281-b951-d872f2087c98