Add gpu_content_tests to GPU bots.
So we can migrate WebGL conformance tests from browser_tests to content_browser_tests.

Note that we only do it for the GPU waterfall in this CL.  Once it lands and works fine, we'll do it for the chromium.webkit GPU bots and GPU trybots.

BUG=111506
TEST=

Review URL: https://chromiumcodereview.appspot.com/11481003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@171809 0039d316-1c4b-4281-b951-d872f2087c98