Allow entries in FILES.cfg to leave arch unspecified.
If a file doesn't specify an arch then it is to be
included for all architectures.   This was motivated
by adding 'arm' to the linux FILES.cfg file.  Rather
than add 'arm' to all of existing files it makes more
sense to allow architecture-all files to simply not
specify an arch list.  This also simplified the FILES.cfg
for non-linux platforms where 'arch' is even more
meaningless.

Also, add a --arch argument to archive_build.py so that
the build architecture can be different to the host
architecture, as is the case with new ARM cross bot.


Review URL: https://chromiumcodereview.appspot.com/11552026

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@172929 0039d316-1c4b-4281-b951-d872f2087c98