Bugfixing URL construction for extract build step
On Windows, if the value for the --build-url argument doesn't include
the filename, auto-generation of the filename ends up adding a backslash
due to use of os.path.join in the construction of the URL. This bugfix
simply changes it to use a forward slash, which should always be correct
for URLs.

BUG=none (but it breaks the ChromiumWebRTC waterfall)
TEST=Successsfully executed extract_build step on a slave with local
modifications to this script.


Review URL: https://chromiumcodereview.appspot.com/11451004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@175507 0039d316-1c4b-4281-b951-d872f2087c98