Fix a poor path assumption in extract_build.
The setup of this code looks like it was a quick fix to correct a bad
--build-dir parameter on Linux (not sure about the other platforms). The bad
value which USED to be passed in was 'src/build'. Now, if someone passes in a
real value, we'll use that instead, and print out a hideous warning if someone
uses the old value.

R=cmp, thestig@chromium.org
BUG=169336

Review URL: https://chromiumcodereview.appspot.com/11820057

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@176188 0039d316-1c4b-4281-b951-d872f2087c98