Add ChromiumWebRTC FYI waterfall
This waterfall has the purpose of providing support for:
* updating the WebRTC stable branch (use status for latest trunk)
* updating webrtc_revision in Chromium DEPS (use status for latest 
  stable).

Triggering of builders:
The idea is to trigger on new revisions in WebRTC trunk and stable 
branch. Since these change quite infrequently compared to Chromium, 
I've also added a Periodic scheduler with 4 hour intervals that 
triggers both the trunk and stable builders.
That way we can narrow down the number of Chromium revisions that 
has passed between a failing build and the latest passing build.
Eventually we would like to add more machines and separate 
builders/testers to bring down the interval even further, but this 
is good enough to start with.

This CL also changes the master.chromium.webrtc waterfall to build 
plain Chromium builds instead of using WebRTC ToT, since that's being 
done in this waterfall now. 
It is moved in here since it should be consdiered a FYI build if the 
latest WebRTC trunk would break Chromium.

Support for Periodic schedulers is added to 
scripts/master/master_config.py

BUG=none
TEST=local builds.


Review URL: https://chromiumcodereview.appspot.com/12033076

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@178817 0039d316-1c4b-4281-b951-d872f2087c98