Allows Zip64 extensions.
According to python's documentation :"ZIP64 extensions are disabled by default because the default zip and unzip commands on Unix (the InfoZIP utilities) dont support these extensions". As we use the python zip only on Windows this is not a problem for us and we should enable the ZIP64 extensions.

BUG=


Review URL: https://chromiumcodereview.appspot.com/12212073

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@182273 0039d316-1c4b-4281-b951-d872f2087c98