Add a ChromiumOS Linux bot to FYI.
It's useful to the sync team (and perhaps others) to be able to
run the browser tests in a ChromiumOS-like environment without
the complexity of setting up a full-fledged ChromiumOS instance.

This bot runs under the FYI bots to act as a staging area to
observe the behavior of the tests without closing the tree should
they be horrendously broken.

BUG=


Review URL: https://chromiumcodereview.appspot.com/12543008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@186512 0039d316-1c4b-4281-b951-d872f2087c98