Fix ChromiumOS Linux slave.
Change ChromiumOS Linux to a BuilderTester.

A bot of type Tester requires a build URL. If you don't supply one,
buildbot gets unhappy; when it tries to transform a list of command line
args to a string, it gets None for one of the values and chokes.

BuilderTester doesn't require that parameter, so let's use that.


Review URL: https://chromiumcodereview.appspot.com/12432008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@186809 0039d316-1c4b-4281-b951-d872f2087c98