Adding master.toolkit skeleton infrastructure.
This is most of the infra necessary to get master.toolkit up and running.
There are two more pieces needed: the toolkit recipe and the toolkit step
generator script. The former will live in our repo (and will get added to this
CL tomorrow) and the latter will live in the toolkit github repo, and I'll
issue a pull request for that soon.

R=iannucci@chromium.org,ykyyip@chromium.org,mmcnulty@chromium.org

BUG=https://code.google.com/p/chromium/issues/detail?id=180003

Review URL: https://codereview.chromium.org/13012002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@190453 0039d316-1c4b-4281-b951-d872f2087c98