Add an "lkgr_type" descriptive field for logging in lkgr finder
Another step to supporting both Blink and Chromium LKGRs; this one
makes the logging a bit clearer so that we can see if this'll be
a Blink LKGR or a Chromium LKGR (although we only actually support
Chromium so far).

This patch also sets some local variables and renames some of the globals
to indicate that they are Chromium-specific.

R=xusydoc@chromium.org
BUG=233418

Review URL: https://chromiumcodereview.appspot.com/13842030

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@196575 0039d316-1c4b-4281-b951-d872f2087c98