add more control over the emails sent in lkgr_finder
This adds command line flags to customize the email address to use
when errors happen, and an email address to optionally use to notify
when the lkgr is updated.

This patch also fixes a divide by zero error that can happen if you run
the script too frequently and no revs change.

With this patch we get the version of lkgr_finder I am currently running
locally to update the Blink LKGR.

R=xusydoc@chromium.org
BUG=233418

Review URL: https://chromiumcodereview.appspot.com/13842031

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@196577 0039d316-1c4b-4281-b951-d872f2087c98