Libyuv try server.
Libyuv is used in many different kind of projects. I've been
maintaining a collection of trybots for it in parallel with
our WebRTC trybots. Now that we've moved our waterfall into
Chrome infrastructure, we need to do the same for libyuv, so
we can retire our old machines.

TEST=locally running master and submitting tryjobs.
BUG=none

Review URL: https://chromiumcodereview.appspot.com/14358033

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@196675 0039d316-1c4b-4281-b951-d872f2087c98