Revert "Rewrite masters_test and masters_cfg_test"
This reverts commit r198275.  Before this CL landed,
a checkout without .bot_password in site_config would
pass presubmit.  After this change, without .bot_password
presubmits would fail.  The CQ does not have .bot_password
in its site_config.  (I even tried adding it manually to
test and the CQ removes it, so not only is that a hacky
approach to put it there manually but it doesn't work.)

I suspect the presubmit wasn't functioning all too well
before if the masters aren't hitting the "I'm missing
.bot_password" case on startup.  If this is a real issue,
a nice solution may be to reland 198275 and also modify
GetBotPassword() to early-return when .bot_password is
not present.

TBR=ilevy@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14727007

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@198331 0039d316-1c4b-4281-b951-d872f2087c98