Add GetTestfilter() to normalise 'testfilter' build property parsing.
This build proeprty can be of various value, depending if it is a try job or
not, if it was triggered directly from buildbot's web ui, Rietveld either via
its web UI or git cl try, via the custom HTTP protocol or via chrome-try.

So this function takes care of all possibility so the "client code" doesn't have
to parse the value.

TBR=iannucci@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/15861021

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@202320 0039d316-1c4b-4281-b951-d872f2087c98