Move code around in factories.
This remove 'chromium_commands depends on swarm_commands'. I figured it was
better to do the code move first as an independent CL, which has no effective
functional change, so that the next change is more obvious.

The reason for this is that the next CL will remove the build property
'swarm_tests', which was an hack. It was set in a function in swarm_commands.py
but is needed in code for CalculateIsolatedSha1s in commands.py. Now the
function that implements the equivalent of 'swarm_tests', i.e. GetSwarmTests(),
is now also in commands.py. Bingo!

TBR=iannucci@chromium.org
BUG=242716

Review URL: https://chromiumcodereview.appspot.com/15875015

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@202357 0039d316-1c4b-4281-b951-d872f2087c98