Remove layout_tests_platform parameter from bot config
This parameter is currently used only on the master_mac_webkit* bots, with a
value of 'chromium-mac'. It is no longer needed because since forking from
WebKit, run-webkit-tests.py can always determine the port from the platform.

The motivation for this change is that the 'chromium-mac' port is about to be
renamed to 'mac'.

BUG=231188
R=dpranke

Review URL: https://chromiumcodereview.appspot.com/18339002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@209472 0039d316-1c4b-4281-b951-d872f2087c98