Revivify master.chromium.git
This change is in preparation for the git migration. It makes
master.chromium.git a virtual mirror of master.chromium, except polling changes
from the src.git repo instead of subversion. This will serve as a proof of
concept that all the main waterfall masters can be switched to Git, as well as
part of the git_branch.py workflow.

R=szager@chromium.org
BUG=253980

Review URL: https://chromiumcodereview.appspot.com/18944002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@210916 0039d316-1c4b-4281-b951-d872f2087c98