Skip a bad revision in lkgr_finder.py.
Our lkgr_finder.py script parses 'got_revision' build
properties out of stored Buildbot build data.  For
one build, got_revision was set to a Git hash rather
than a SVN revision.  The SVN rev comparator assumes
that revisions are base-10 numbers and so the script
was crashing when it encountered this failure.

I hardcoded the script to skip the revision in question
so the LKGR value will get updated.  This is meant as
a temporary workaround, though there will need to either
be an immediate removal of the offending got_revision/
build data, or a phased approach to remove this change
once that build data is removed from the Buildbot
master.

TBR=szager@chromium.org
BUG=260368

Review URL: https://chromiumcodereview.appspot.com/19282002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@211688 0039d316-1c4b-4281-b951-d872f2087c98