Prefer swarm_get_results in src/tools/swarm_client.
This is to transition away from using build/third_party/swarm_client into a
properly versionned src/tools/swarm_client/; build/third_party/swarm_client is
not pinned so it uses whatever HEAD is, which is problematic.

After this CL, the dependency to checkout build/third_party/swarm_client will be
removed, then the fallback will be removed once it's verified to work. Doing
it in 3 phases with a warning reduces the likelihood of breaking the build step
along the transition.

R=iannucci@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/19196002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@211827 0039d316-1c4b-4281-b951-d872f2087c98