Fix archive build for git.
Even though there were already nice methods for getting SVN or Git revisions,
archive_build was only using the SVN one, causing master.chromium.git to fail.

This is a resubmission of https://codereview.chromium.org/19059003/, because it
was reverted in https://codereview.chromium.org/19067005/. This incorporates a
new change to the way GitOrSvn determines if it is in a Git or Svn repo which
allows it to work for git repos without breaking svn repos like the first one
did.

R=szager@chromium.org,maruel@chromium.org

Review URL: https://chromiumcodereview.appspot.com/20149005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@213664 0039d316-1c4b-4281-b951-d872f2087c98