Try again to fix 'archive_webkit_test_results'.
Turns out it wasn't just a typo, and so r214134 didn't
work, either. This change patches things so we expand
buildername and buildnumber before running the step, so
that we don't need buildername or buildnumber to be passed
down and interpolated in the log parser.

R=iannucci@chromium.org
BUG=261732

Review URL: https://chromiumcodereview.appspot.com/20969003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@214140 0039d316-1c4b-4281-b951-d872f2087c98