Improve the blink_trybot testing code.
Refactors the big blob of json to a couple functions to procedurally generate it.
Now that is_unxepected is calculated dynamically (aka 'correctly') the
expectations reflect this.

This is in order to support the upcoming patch to try the reduced set of webkit
tests as the initial second pass.

R=dpranke@chromium.org
BUG=265271

Review URL: https://chromiumcodereview.appspot.com/22876003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@217324 0039d316-1c4b-4281-b951-d872f2087c98