Unshard WebKit Win7 dbg bots.
With the recent speed increases in debug builds of content_shell,
and the move to a new VM on the 720xd's, there's not really a good
reason to run the debug bots sharded any more. In the future,
if we need to shard tests across multiple machines, we'll want to
use swarming and isolates anyway.

R=ojan@chromium.org, iannucci@chromium.org, pschmidt@chromium.org
BUG=270323

Review URL: https://chromiumcodereview.appspot.com/22928006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@217370 0039d316-1c4b-4281-b951-d872f2087c98