Add check for virtual webcam on bots running WebRTC tests.
We've had lots of issues with the virtual webcams not being present on
bots in the past. This CL adds a step that runs a script that ensures
the required virtual webcam software is started before running tests.

One may think that rebooting the testers between each build would be a
better solution, but having this will also catch other reasons for
failure, like when:
- the software is not yet setup on the bot.
- Linux kernel updates requires a recompile of the kernel module.
- permission problems cause the virtual webcam software to fail.

Implementation note:
During testing of the script as a buildbot step, I discovered that
the v4l2_file_player is automatically killed as soon as the step
finishes unless usePTY=False is specified and stdout/stderr/stdin is
redirected to subprocess.PIPE.

TEST=successful tests running local running master and Linux, Mac and Windows slaves.
BUG=none

Review URL: https://chromiumcodereview.appspot.com/14631018

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@217503 0039d316-1c4b-4281-b951-d872f2087c98