V8 Buildbot: Prevent gclient from using revision flag in order to enable safe sync.
The revision string from the scheduler (e.g. branches/bleeding_edge@16183) is only used by gclient if the part before @ matches a gclient solution name. This is never the case on the V8 waterfall since standalone solutions are called "v8" and chromium solutions "src". That emits a warning in update steps on all builders: "Please fix your script...".

The corrupt revision is thus never used, which is no problem as such, since the gclient spec contains all that is needed.

When used together with safe sync, however, the --revision flag disables the safe sync feature, even though it is not used in the end.

The no_gclient_revision property disables setting the revision flag.

BUG=272090

Review URL: https://chromiumcodereview.appspot.com/23167003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@217537 0039d316-1c4b-4281-b951-d872f2087c98