self.command may not have .stdout for certain commands.
It's a problem because self.command is constantly overriden, instead of
accumulating them. So change for parseGotRevision() to be called
potentially twice. It's not a problem because
scripts/master/chromium_step.py:252 takes the last value when multiple
values for a property are sent. I.e.:
  got_revision = cmd.updates[cmd_arg][-1]

This issue only happened on nacl slaves, which I hadn't tested.

R=phajdan.jr@chromium.org
BUG=273215

Review URL: https://chromiumcodereview.appspot.com/22902004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@217684 0039d316-1c4b-4281-b951-d872f2087c98