Make it possible to specify TSan v2 suppressions file.
In WebRTC, we're aiming at replacing our TSan bot with TSan v2.
To do this, we have to be able to specify and maintain our own
suppressions file to be used with our tests.

This CL makes it possible to specify the path to the suppressions
file using a factory property ('tsan_suppressions_file') for the
builder instead of having it being hardcoded in runtest.py.

BUG=274414
TEST=none

Review URL: https://chromiumcodereview.appspot.com/23143005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@218061 0039d316-1c4b-4281-b951-d872f2087c98