Retire libjingle master and move tests to client.webrtc.
Also add the libjingle tests to the WebRTC trybots, making them
useful for trying out libjingle fixes.

All the Linux slave machines will have to have Sun Java 6
installed, but as this is the default for 
install-build-deps-android.sh it should already be present
on the machines.

BUG=275374
TEST=Ran a succesful build on one of the Linux builders.

Review URL: https://chromiumcodereview.appspot.com/23151007

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@218421 0039d316-1c4b-4281-b951-d872f2087c98