Reapply r217684 and r217626.
Found the bug and tested it locally.

Fix a bug in got_XX_revision when a DEPS roll is involved with git-try/gcl try.
    
The function _handleGotRevision() is called *before* the patch is applied, so it
was unconditionally getting the stale revision. This CL fixes the problem for
try jobs triggered with gcl try or git try, it doesn't fix the issue with git cl
try or Rietveld triggered try jobs.
    
Add unit test. I wasn't sure if this part of the code was working for a new
repository I'm fetching so I unittested its output.

R=phajdan.jr@chromium.org
BUG= 273215

Review URL: https://chromiumcodereview.appspot.com/23228003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@218526 0039d316-1c4b-4281-b951-d872f2087c98