WebRTC FYI waterfall
A waterfall for non-breaking tests for WebRTC. This will be a staging
area for new bots that are not yet ready to run in the official
waterfall or that should be FYI for other reasons.

This CL will only add Linux TSan v2 and TSan RV bots for now,
but the CL prepares for Dr Memory bots on Windows as well.

BUG=274414
TEST=Ran successful builds with local master and slave.

Review URL: https://chromiumcodereview.appspot.com/23231004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@219777 0039d316-1c4b-4281-b951-d872f2087c98