Add a new buildbot configuration for the Push Messaging Canary Test
The Push Messaging canary test buildbot should not be enabled for the waterfall.
This test uses the network, so it will be flaky with network conditions.  The
Chrome Apps team uses the test to monitor the push messsaging servers and to
see if checkins have broken the Push Messaging feature.

The test works by compiling the latest browser_tests executable, and
using an extension to send a push message, and recieving the push
message after a round trip.  The test extension is already checked in,
this patch is just the buildbot test framework to run the existing test.

Review URL: https://chromiumcodereview.appspot.com/16901006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@219874 0039d316-1c4b-4281-b951-d872f2087c98