Fix the link to the archived webkit results on the bots.
Fixing the link required modifying the way we initialize the
log_parser for the archive logs so that we know not to try and
include the revision number in the text that is output for the link.

This change is another attempt to fix the issue that I had tried
to fix in r212503, r214134, r214140 . This time I actually know
what's going on and have tested it locally to confirm that it works :).

R=xusydoc@chromium.org, iannucci@chromium.org, ilevy@chromium.org
BUG=261723

Review URL: https://chromiumcodereview.appspot.com/23670003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@220996 0039d316-1c4b-4281-b951-d872f2087c98