update links for archived layout test results
Now that the archive_layout_tests_wrapper.py is archiving both the full
directory and the zip files, we can link directly to the results.html
and not force users to download the zip to see the results. This patch
changes the recipe to display both the results.html link and the .zip link.

R=iannucci@chromium.org
BUG=261719

Review URL: https://chromiumcodereview.appspot.com/24002002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@221828 0039d316-1c4b-4281-b951-d872f2087c98