Increase minimum buildcache size to 30.
This is an unintended consequence of autoBuildCacheRatio. For builders with a small number of slaves, a buildCache minimum is imposed. That minimum is less than the typical size of the buildbot console. This means that each console access triggers several cache misses on all builders. The console also uses a suboptimal access pattern which can cause needless cache eviction.

BUG=277508
TBR=maruel@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23651011

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@223325 0039d316-1c4b-4281-b951-d872f2087c98