Change the DISABLED_ test to a MANUAL_ test to better fit the description.
The Push Messaging Canary test is normally run manually only, and should never
be run in the commit queue or automatically since it depends on the network,
and will be flaky when the network has long response times.  We had it marked
as DISABLED_ before to keep it from being run automatically, but we discovered
that MANUAL_ was a better reflection of what we actually do with this test.
The test itself has already been changed, but this buildbot needs to be updated
to reflect the test changes.

Also, I tried to choose better ports for running the tests.

bug=284828

Review URL: https://chromiumcodereview.appspot.com/23757046

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@223646 0039d316-1c4b-4281-b951-d872f2087c98