Add an AutoHide value for config's hidden.
It will cause the item to be hidden if it's in the 'default' state (i.e. its data
is the empty_val), and non-hidden otherwise. This cuts out a lot of cruft from
the various expectations, as well as no longer omitting non-critical values in
real recipe runs.

R=agable@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/23536073

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@224234 0039d316-1c4b-4281-b951-d872f2087c98