Fix step name problem for endure tests.
Currently the stdout for endure cannot be displayed on the waterfall due to the slash in the name.

This CL keeps the slash for the hierarchical naming on the dashboard, while using underscore for the step name.

There is also a bit clean up of repetitive code.

Review URL: https://chromiumcodereview.appspot.com/24247005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@224360 0039d316-1c4b-4281-b951-d872f2087c98