Add libjingle_peerconnection_objc_test to Mac x64 bots.
The use of the 'custom_cmd_line_tests' factory property is needed
since the test is executed from the
https://code.google.com/p/webrtc/source/browse/trunk/webrtc/test/buildbot_tests.py#41 helper script.

The mac_sdk=10.7 is needed since the target is not
generated otherwise. The Mac VMs have the 10.6 SDK
present, which will be picked unless another SDK is
specified (see Chromium's common.gypi).

TEST=local build successful. Also tested building and
running the test on a Mac VM in Chrome infra.
BUG=webrtc:2120

Review URL: https://chromiumcodereview.appspot.com/23620067

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@225059 0039d316-1c4b-4281-b951-d872f2087c98