Add new NaCl SDK slave that does nothing but build the SDK.
The idea here is that this trybot will be fast and
reliable enough to be part of the CQ and have and
equivalent tree-closer on the main waterfall.

Currently this bot does not build chrome at all and doesn't
run any tests that rely on the browser.  The aim is to have
fast, tree-closing, builder that can catch SDK build breaks.

This change relies on:
https://codereview.chromium.org/25597003

TEST=Linux bot tested locally
BUG=302579
R=binji@chromium.org, iannucci@chromium.org

Review URL: https://codereview.chromium.org/25586002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@226585 0039d316-1c4b-4281-b951-d872f2087c98