Turn off ash and aura by default on linux blink bots.
We're getting ready to enable Aura and Ash on Linux by default, but the
layout tests need to be updated to use the mock theme. Until then,
we'll force Aura and Ash off on the bots so that the tests continue to pass.

(This is the same as we're currently doing on the Win bots).

R=iannucci@chromium.org
BUG=125106

Review URL: https://chromiumcodereview.appspot.com/25792002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@226850 0039d316-1c4b-4281-b951-d872f2087c98