Fixed problems in parsing of gtest output for flakiness dashboard.
Most harnesses which upload results to the flakiness dashboard use the
--results-directory argument to runtest.py in order to produce an XML
file which is transformed to JSON and uploaded. If just the
--generate-json-file argument is passed then the gtest output is
supposed to be parsed.

The runtest.py change fixes an exception thrown within runtest if
--generate-json-file is passed without --results-directory.

The gtest_slave_utils.py change fixes an incorrect short-circuit in
the fetching of the results map.

These changes have been tested locally but not yet on a waterfall, so
it isn't known yet if they'll fix all the issues.

BUG=225939
R=stip@chromium.org
TBR=stip@chromium.org

Review URL: https://chromiumcodereview.appspot.com/26083003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@227153 0039d316-1c4b-4281-b951-d872f2087c98