Move memory waterfall build directory computation to the slave.
runtest.py already jumps through the right hoops to get the right build_dir
(mostly; but it'll have to learn how to do this completely correct soon),
so just give that an option to pass the build directory on to the memory
scripts.

With this, the isolate master is the last place to call ConvertBuildDirToLegacy
on the master.

Depends on https://codereview.chromium.org/26190002/

BUG=294387
Review URL: https://codereview.chromium.org/26184003/

(reuploading to land)
TBR=maruel@chromium.org (who reviewed at the above review url)

Review URL: https://chromiumcodereview.appspot.com/26258004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@227327 0039d316-1c4b-4281-b951-d872f2087c98