Add app_list_unittests as a test step and tree closer on mac, win, and linux_aura.
app_list_unittests has been built on all non-linux targets since r178260. It was
added as a default build target for linux_aura in r225936.

On linux_chromeos (only) it is already a test step in the waterfall builders, a tree
closer, and a CQ step for linux_chromeos.

This CL adds app_list_unittests as a test step and tree closer on all platforms
that build the target, which is now everything but gtk-linux and mobile.

BUG=169114, 166730
R=stip@chromium.org, xusydoc@chromium.org

Review URL: https://codereview.chromium.org/27296002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@229248 0039d316-1c4b-4281-b951-d872f2087c98