Remove support for "graceful shutdown" command in buildbot.
This change is implemented by using an explicit Authz instance
that has a whitelist of allowed commands, which does not allow
graceful shutdown.

Furthermore, as an additional protection, code that supports
the shutdown hooks has been removed from buildbot. The change to do that
is minimal to keep it maintainable. Removing all the code that has to do
with graceful shutdown would touch many more files.

BUG=309046
R=cmp@chromium.org

Review URL: https://codereview.chromium.org/29153004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@229450 0039d316-1c4b-4281-b951-d872f2087c98