compile.py/mac: Translate xcodebuild parameters into ninja parameters.
I'd like to be able to switch to ninja (and back, if needed) without master
restarts. For this, compile.py needs to be able to call ninja even if the master
passes xcodebuild parameters like -project, -target, or -sdk.

-sdk is only used on iOS, so it doesn't need to be supported. -project passes
the xcodeproj file a build target is in, so just ignore that. -target passes
the build target, which ninja does need.

(This is temporary and can be removed once all branches ship using ninja.)

BUG=294387

Review URL: https://codereview.chromium.org/41263002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@230808 0039d316-1c4b-4281-b951-d872f2087c98