Added GLib/DBus workaround to runtest.py under --do-dbus-workaround option.
It's likely we'll want to do this on all the bots and just remove
this command line argument altogether, but I'm concerned about
destabilizing the existing bots that don't reboot between each run.
For the moment, use this argument only on the GPU bots.

It's known to be needed generally, and not specifically for Telemetry
tests, due to failures like this one in content_gl_tests:
http://build.chromium.org/p/chromium.webkit/builders/GPU%20Linux%20%28dbg%29%20%28NVIDIA%29/builds/20405/steps/content_gl_tests/logs/stdio

BUG=309093

Review URL: https://codereview.chromium.org/44923002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@231085 0039d316-1c4b-4281-b951-d872f2087c98