Rewrite paths in .isolated files to be consistent with build_directory.
Builders and testers get their paths from build_directory, based on various
criteria. Because of this, the builder and tester might not have the same
build directory, and the paths in the isolated file might not refer to the
directory the build artifacts were unzipped to.

Example: Say a builder builds to src/build, then zips that up. The isolated file
is created at build time, so it will refer to src/build. The tester might then
unzip it to src/out (if it didn't run runhooks, since src/out is the default if
no recent build/all.sln is present), and isolate tests would then run a stale
binary from src/build.

To fix this, make runisolatedtest.py rewrite build paths in the isolated file
at build time. This is pretty ugly, but it's hopefully only temporary until
all isolate bots use ninja.

While here, also log the command that is about to be run, as the current
output is very misleading.

BUG=311625,311625

Review URL: https://codereview.chromium.org/45623002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@231343 0039d316-1c4b-4281-b951-d872f2087c98