Run create_profiles before archive_build.
It generates files that are supposed to be zipped up and shipped to testers.
As is, it zips up files from the last build.

Also rename it from "Test" to "Step" as this step doesn't test anything.
This addresses item 1 from comment 0 on the linked bug.

BUG=313317

Review URL: https://codereview.chromium.org/53683002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@232128 0039d316-1c4b-4281-b951-d872f2087c98