Added GPU builders to chromium.gpu, chromium.gpu.fyi and chromium.webkit waterfalls.
Moved recipe and build_config into the slave.cfg (except on the
chromium.webkit waterfall) to simplify setup logic.

This is the first of two steps to cut over these waterfalls to a
builder/tester model. The second step will be to switch the remaining
GPU bots on these waterfalls to use the download_and_test recipe, once
the builders are known to be working.

Tested locally with scripts/slave/runbuild.py's --list-builders and
--list-steps commands.

BUG=310926

Review URL: https://codereview.chromium.org/68533010

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@234392 0039d316-1c4b-4281-b951-d872f2087c98