Revert "Replace Linux Tests with Linux Aura as a Tester."
This reverts commit cb3ef08d7371cc51ed00a67ef48d2828772cada9.

This commit was part of an effort to get Linux Aura on ToT to test quickly. It
kinda worked, but we don't have good evidence as to *why*. Building Aura on ToT
by default is being reverted, so we are reverting this CL as well.

BUG=319075
R=cmp@chromium.org, dpranke@chromium.org, erg@chromium.org

Review URL: https://codereview.chromium.org/65713005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@235271 0039d316-1c4b-4281-b951-d872f2087c98