Stop forcing Aura off on Blink bots.
As we now support running the layout tests under Aura, we should just use
the default settings as configured in build/common.gypi. Currently this
would mean that we'd start running the tests under Aura on Windows, but
not on Linux.

This effectively reverts r226850, r223160, and r218587.

R=iannucci@chromium.org
BUG=229207, 318980

Review URL: https://codereview.chromium.org/75603002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@235830 0039d316-1c4b-4281-b951-d872f2087c98