Respect parent_got_revision in gclient module.
This comes into play for triggered builds which should match the triggering
build's gclient configuration as closely as possible. Unfortunately, the
got_*_revision 'solution' is pretty hacky, so we do our best to be compatible
here.

In The Future, we would like to pin the entire manifest on the parent builder
(or better, at the time we trigger the job).

R=kbr@chromium.org, stip@chromium.org
BUG=310926

Review URL: https://codereview.chromium.org/61583005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@235902 0039d316-1c4b-4281-b951-d872f2087c98