Patch path filtering script.
A script that makes it possible to filter out files from a patch
using paths, which can be useful for some trybot situations.

It is used if additional flags are passed to the apply_svn_patch.py script.
This is required in order to make https://codereview.chromium.org/75163006/
possible.

TEST=local testing and unit tests.
BUG=312827

Review URL: https://codereview.chromium.org/27575002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@240099 0039d316-1c4b-4281-b951-d872f2087c98