Remove passing of the --root-dir to the filter script
When absolute paths are passed as the --root-dir value,
it becomes pretty useless in it's current implementation
since it becomes impossible to filter the right paths.

This CL removes the use of it since it turned out not
being needed for my current use case anyway.
(it was added in my initial CL at
https://codereview.chromium.org/27575002/ in belief it
would be needed).

This is also a bit cleaner since we no longer put any flag
requirements on the filter script.

BUG=312827
TEST=unit tests and local testing with the 'patch' command line in apply_svn_patch.py replaced with 'cat' and running like this: ./apply_svn_patch.py -p file:///path/to/local/svn/repo/git_patch2.patch -r . --filter-script=patch_path_filter.py -- --path-filter=chrome/
R=iannucci@chromium.org

Review URL: https://codereview.chromium.org/114983006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@240780 0039d316-1c4b-4281-b951-d872f2087c98