Utilize slavebuilddir to save space and improve sync speed.
The slavebuilddir property is already used in other places,
mainly in master.tryserver.chromium.
It allows to specify the name of the slave directory so that
multiple builders can share the same checkout. This is safe to
do when using ninja as all the project files and build output
is put in Debug/Release in out/.
Some other cases have unique directories as well (Win x64 and iOS)

As we share machines between builders in WebRTC, this will save
significant amount of space on the slaves.

TEST=Verified the right dir is used on a local master+slave.
BUG=none

Review URL: https://codereview.chromium.org/106923004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@241525 0039d316-1c4b-4281-b951-d872f2087c98