Remove support for "graceful shutdown" command in buildbot.
This change is implemented by using an explicit Authz instance
that has a whitelist of allowed commands, which does not allow
graceful shutdown.

Furthermore, as an additional protection, code that supports
the shutdown hooks has been removed from buildbot. The change to do that
is minimal to keep it maintainable. Removing all the code that has to do
with graceful shutdown would touch many more files.

This was originally uploaded as https://codereview.chromium.org/29153004/

BUG=309046

Review URL: https://codereview.chromium.org/118173002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@243084 0039d316-1c4b-4281-b951-d872f2087c98