Add 'hostname' as a dimension.
This permits running a job on a specific slave.

This is only meant for administration purposes, for example cleaning up every
slave, one at a time. The advantage of using swarming itself to run these is
that they will be run in-band, so no job will have to be canceled to do this.
The scheduler will simply select the task according to the priority, which will
likely be either very high or very low.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.chromium.org/131293003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@244899 0039d316-1c4b-4281-b951-d872f2087c98