Support component revisions in cf archive script.
This allows using the "revision_dir" factory property just like in zip_build.py. If set, the revision dir of the component will be used to retrieve the revision.

Example:
- revision_dir = "v8"
- V8 revision will be retrieved from src/v8
- Zip will be named <name>-v8-component-<v8 revision>.zip

If "revision_dir" is not set, there is no intended change, except (unintended change): because of the generality of slave_utils.GetBuildRevisions, in case of a git checkout, a git hash would be returned, while the old function would have just failed. Since there are svn checkouts on the bots, it shouldn't make a difference.

BUG=282230

Review URL: https://codereview.chromium.org/140753002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@245278 0039d316-1c4b-4281-b951-d872f2087c98