Cleanup sizes.py script and update lib path for WebView chromium.
After https://chromereviews.googleplex.com/21617018/, the WebView
telemetry bot is symlinking the .so in the chromium out directory.
A prior change to sizes.py (crrev.com/57803006), in fact, has
required a refactoring of the WebView telemetry bot recipe, since
it was making infeasible referencing the .so from outside the
chromium checkout.
At this point, however, it looks like the --target-dir argument
is not used anymore inside sizes.py. Hence this CL is turning it
into an ignored argument.

BUG=314253

Review URL: https://codereview.chromium.org/140743005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@245366 0039d316-1c4b-4281-b951-d872f2087c98