Cherry picking "Doing a shuffle of ios slaves so that I can proceed with the bug below."
The slaves all have the latest Xcode with both new and old (which was the problem the first time) SDKs. Manual compilations succeed on build80-a4. This should be ready whenever the tryserver is ready for it.

This CL was submitted prematurely, reverted, and re-reverted many times. We still need the original CL, so I'm cherry picking it to avoid a long "Revert of" chain.

Original description:
Doing a shuffle of ios slaves so that I can proceed with the bug below.

BUG=325995

Review URL: https://codereview.chromium.org/98173010

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@242717 0039d316-1c4b-4281-b951-d872f2087c98

Review URL: https://codereview.chromium.org/130503006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@246214 0039d316-1c4b-4281-b951-d872f2087c98