Update GPU recipes' test expectations in preparation for switch to isolates.
This CL has the following functional changes:

1) Factor out the names of the GPU isolates into a file shared between
api.py and test_api.py.

2) Change the isolates' hashing scheme in the test expectations for
clarity.

3) Make all of the isolates' hashes available to the GPU recipe tests
which will soon need them.

4) Switch to invoking run_gpu_test.py instead of run_gpu_test, so that
the latter can be deleted.

5) Added a skip_checkout flag to the gpu/download_and_test recipe for
local testing purposes.

The purpose of this CL is to reduce the size of the test expectations
update in a subsequent CL which will switch to running the tests via
isolates.

BUG=321878

Review URL: https://codereview.chromium.org/132233048

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@246471 0039d316-1c4b-4281-b951-d872f2087c98