Add runhooks for WebRTC Android APK Tester builders.
The runhooks step was previously removed for the APK builders
and testers in https://codereview.chromium.org/85243005/

However, while the builders executes runhooks as a part of the compile step, the testers don't have a compile step.
Because of this, they must run it as a part of the update step.

TEST=none
BUG=none

Review URL: https://codereview.chromium.org/131513009

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@246839 0039d316-1c4b-4281-b951-d872f2087c98