Add LKCR (Last Good Compilable Revision) to LKGR Finder
Note that the amount of builders it watches doesn't really make a difference
in a typical case where most of the things compile. On the other hand, once
in a while we get an odd compile failure that is ChromeOS-specific
or iOS specific or only happens in Debug mode on Linux (e.g. because of
_GLIBCXX_DEBUG or some other code only compiled in Debug).

LKCR should really mean Compilable, not Maybe Compilable or Mostly Compilable.

Also, include runhooks because I've seen them failing with weird errors.
In theory it should halt the build, but just to be safe and future-proof
I think it's better to include them.

BUG=317931

Review URL: https://codereview.chromium.org/138063008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@246985 0039d316-1c4b-4281-b951-d872f2087c98