Try to work around CRLF on Windows.
In particular, do not use universal_newlines=True so the buffering is not
affected.

This only affects the canary master since it is the only one which has Windows
slaves that runs swarming steps.

I'm not 100% this fix will work but we'll see quickly enough on the Canary
master.

TBR=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/148183006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@247233 0039d316-1c4b-4281-b951-d872f2087c98