chromium_trybot: only fall back to LKCR when compile fails
This is intended to address apply_issue failures happening when patch
is based on more recent revision than LKCR and does not apply cleanly
on earlier revisions.

Note that this cannot make things worse: if compile fails, we'll fall back
to LKCR anyway. Otherwise the patch has a higher chance of applying,
and if it doesn't apply cleanly on HEAD it won't be committed anyway.

BUG=317931

Review URL: https://codereview.chromium.org/167923002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@251912 0039d316-1c4b-4281-b951-d872f2087c98