Implicitly pass builder name, slave name, and master name to runtest.py.
All builders on all waterfalls specify these values, so stop passing
them explicitly to api.chromium.runtest. Fix up recipe tests which
weren't defining these properties.

A couple of recipes explicitly test that the recipe works if a build
number and revision aren't passed. Query for the build number and pass
it if specified in the build properties.

This is a prerequisite cleanup for a follow-on CL which will remove
the src/ checkout for the bulk of the GPU bots.

BUG=321878
R=iannucci@chromium.org

Review URL: https://codereview.chromium.org/168263002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@252140 0039d316-1c4b-4281-b951-d872f2087c98