Rename triggered_builder to swarming_triggered_builder to be clearer.
It will become the signal to know if the swarming specific steps should be run
or not so prefix it with 'swarming_' so its purpose is clear.

Stop passing the predefined list of tests to run on swarming to
manifest_to_hash.py, the script has been ignoring the values since r251882.

Remove last use_swarm_client_revision which was forgotten from r252075.

Triggers the swarming dependent builder a tad sooner.

R=stip@chromium.org
BUG=

Review URL: https://codereview.chromium.org/172863006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@252202 0039d316-1c4b-4281-b951-d872f2087c98