Fix broken uploading of flakiness dashboard results from recipes.
Under Issue 321878 I began passing a new --master-name argument to
runtest.py from recipes, mainly to allow it to be mocked for local
testing of recipes. In doing so, the wrong value was being passed to
the flakiness dashboard when uploading results. This silently broke
uploading of flakiness dashboard results for all recipe based bots.

Stop passing this argument to runtest.py from recipes 99% of the time.

Rename the argument to runtest.py to --master-class-name to clarify
what it really is.

Allow mocking of this value first in the GPU recipes. Other recipes
can start doing so later if desired.

BUG=347216
R=iannucci@chromium.org, stip@chromium.org

Review URL: https://codereview.chromium.org/181653008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@253895 0039d316-1c4b-4281-b951-d872f2087c98