Fix the Chromium revisions on the flakiness dashboard for the Blink GPU bots.
The Blink GPU bots are alternately triggered by Blink and Chromium changes.
Therefore the 'revision' build property is not a good value to use. Instead
use parent_got_revision, which is available on all of the test bots, and
fall back to querying the workspace, as the code for the WebKit revision
already does. Developers using the build_and_test recipe for local testing
can manually specify parent_got_revision if they are skipping the checkout.

BUG=347726

Review URL: https://codereview.chromium.org/185073002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@254273 0039d316-1c4b-4281-b951-d872f2087c98