Configure "iOS Device (ninja)" to build fat binaries on fyi bots
To prevent regression we want at least one of the iOS bot to build the code
for both armv7 and arm64. Since the support for fat binaries is developped
primarily for ninja build system, enable it for those bots.

Note: the "target_subarch" gyp variable is control whether the build should
be for armv7 (the default), arm64 or both (fat binary).

BUG=348001

Review URL: https://codereview.chromium.org/176993009

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@254739 0039d316-1c4b-4281-b951-d872f2087c98