Remove iOS builders WebRTC gatekeeper (take 2)
The changes in https://codereview.chromium.org/203543006/
was not enough to exclude the ios bots from the gatekeeper.
Changing the category to None "must" be fixing this (I made
the previous change with the assumption that a string that is
not present in master_gatekeeper_cfg.py would make the builder
not being included, which was wrong).

BUG=webrtc:3073
TEST=None

Review URL: https://codereview.chromium.org/210903002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@259182 0039d316-1c4b-4281-b951-d872f2087c98