WebRTC: Enable Goma for Windows builders.
Use ChromiumFactory's helper methods for setting up the
GYP environment correctly.
Having that, we are able to compile using Goma on Windows,
which is turned on here.

All desktop platforms got Goma turned on in
https://codereview.chromium.org/197873038/
but I had to revert on Windows in 
https://codereview.chromium.org/208683005/
due to compile breakage.

BUG=None
TEST=Ran a successful local Debug and Release build with a Windows slave.

Review URL: https://codereview.chromium.org/223313002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@261326 0039d316-1c4b-4281-b951-d872f2087c98