10% experiment on removing NO_NACL_GOMA flag.
In the previous config, we have tried removing NO_NACL_GOMA from selected
hosts on Linux, Mac, and Win.  It looks not causing large issues.

To move this forward, let me apply this to "randomly" selected host.
To make it easy to trace the broken host on this experiment, the selection
is done with the value calculated from its hostname.

Review URL: https://codereview.chromium.org/223143005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@261580 0039d316-1c4b-4281-b951-d872f2087c98