Create a gsutil api function that takes a URL
GSUtilApi has an upload step that gives an https:// URL to the upload.
It also has a download step that accepts a gs:// URL to the file.  It
is convenient to also be able to download with the https:// URL provided
by the upload step.

BUG=358659
R=agable@chromium.org, mrapisarda@google.com

Review URL: https://codereview.chromium.org/221313002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@261830 0039d316-1c4b-4281-b951-d872f2087c98