WebRTC: Switch client.webrtc Android bots over to recipes.
Fix the recipe so the got_revision reported to buildbot
comes from the src/third_party/webrtc dir instead of the
Chromium revision in src.

The perf configuration for the testers have never been used
so far, since we were blocked by lack of support for arbitrary
tests logging perf numbers in the Chrome for Android toolchain
scripts. It's just one test affected in WebRTC and the plan is
to move that into a separate executable to run on the chromium.perf
waterfall anyway.

BUG=346198
TEST=

Review URL: https://codereview.chromium.org/229463005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@262688 0039d316-1c4b-4281-b951-d872f2087c98