Make results_dashboard send just one request per test run (version 2).
This is a second version of crrev.com/217053012.

The previous version got reverted because it appeared to be causing a couple of errors:
 1. Undefined variable "revision" (probably because of empty logs_dict)
 2. ValueError when calling results['supplemental_columns'].update(supplemental_columns), probably because the supplemental_columns passed in wasn't a dictionary.

It seems like both of these might be caused by some incorrect values are passed into SendResults, but the arguments passed to SendResults from runtest.py seem to match the ones expected in results_dashboard.py.

Does this look like the right way to validate arguments?
Should I leave off the STEP_EXCEPTION part?

BUG=

Review URL: https://codereview.chromium.org/227233011

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@262749 0039d316-1c4b-4281-b951-d872f2087c98