WebRTC: Add Dr Memory "Full" trybot
In https://codereview.chromium.org/225663005/
a Dr Memory trybot was added for the "Light" configuration.
The idea behind this was that the "Full" configuration takes
too long to execute for a trybot (~3 hours).
Considering that some developers still don't have Windows
machines and that you can pass test filters to the trybots,
I think it's useful to have a trybot for the "Full" configuration
that is not part of the default set of trybots (decided
by the list in the PRESUBMIT.py of WebRTC checkouts).

BUG=360054

Review URL: https://codereview.chromium.org/232313003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@263607 0039d316-1c4b-4281-b951-d872f2087c98