Make results_dashboard send just one request per test run. (version 3)
After version 2, https://codereview.chromium.org/227233011/, we verified that somehow the old list of arguments are being passed to SendResults somewhere somehow even though that CL was also supposed to change the call site in runtest.py.

For this CL, I've made it so that SendResults would also work if called with the original list of arguments; it takes an optional parameter logs_dict, and if logs_dict wasn't given it makes it.
BUG=

Review URL: https://codereview.chromium.org/236593004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@264828 0039d316-1c4b-4281-b951-d872f2087c98