chromium_win_clang: run the standard set of windows tests
This seems to be tests_win_1 + tests_win_2. I noticed that this doesn't
include content_unittests, which we're currently running. Maybe that should
be added to this list?

Also, don't set Clobber=true. The clang update script will automatically clobber
the build if a new clang version is pulled down and it would be good for the
bot the exercise this functionality.

BUG=82385

Review URL: https://codereview.chromium.org/253343002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@265944 0039d316-1c4b-4281-b951-d872f2087c98