Handle other "passing" results from run-webkit-tests properly.
We were rejecting a test result as a failure because the results from
run-webkit-tests were ['CRASH', 'NEEDSMANUALREBASELINE']; the latter
result indicates that the test actually passed.

I really need to re-work run-webkit-tests to just spit out the list
of failures that matter rather than force the recipe to duplicate all of this
complicated logic.

R=iannucci@chromium.org, eseidel@chromium.org, ojan@chromium.org
BUG=366439

Review URL: https://codereview.chromium.org/250743002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@266035 0039d316-1c4b-4281-b951-d872f2087c98