Added TryJobGerritScheduler
The TryJobGerritScheduler uses TryJobGerritPoller to poll open issues on
Gerrit with TryJob=1 label. For each issue it searches for a 'new' message in
that starts with !tryjob prefix. A 'new' message means its date is
greater than previous poll time.

If !tryjob is followed by json, it is parsed as a job description. The
format is {'builderNames': <list of strings>}. The short form for this
is just a list, e.g. ['Stable'].

For each try job, TryJobGerritScheduler/TryJobGerritPoller create
1) a new buildbot change with the Gerrit patchset revision
2) a Sourcestamp
3) a Buildset with specified builder names. If not specified, the
default builder list is used.

The TryJobGerritPoller is a part of TryJobGerritScheduler and not
designed to be used otherwise.

This was tested locally with the following CL as input:
https://quickoffice-internal-review.googlesource.com/#/c/11081/1

BUG=366099, 364639, 364631
R=kmg@chromium.org, vadimsh@chromium.org, dnj@chromium.org

Review URL: https://codereview.chromium.org/250983003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@266962 0039d316-1c4b-4281-b951-d872f2087c98