Remove 'WebKit Linux Leak (dbg)' and add 'WebKit Linux Oilpan Leak'
We found that the release version of the leak detection bot was enough,
so we'd like to remove that.

At the same time, we'd like to add a new leak detection bot for Oilpan,
which we expect detects the leaks caused by Oilpan.

stip: As for the VM, I wonder it is possible to divert the VM of 'WebKit
Linux Leak (dbg)' to 'WebKit Linux Oilpan Leak'. I'll create a ticket
if needed.

BUG=359024

Review URL: https://codereview.chromium.org/250533002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@267024 0039d316-1c4b-4281-b951-d872f2087c98