WebRTC recipe: Set parent_got_revision mapping for testers.
In the last deployment attempt for the Android APK builders+testers,
it was discovered that the parent_got_revision property results in
the testers trying to sync --src@revision instead of --src@HEAD which
is the intention with these builders (the WebRTC revision being polled
is the target for the testing, the Chrome checkout is there only to
make it possible to build the tests using the same test framework).

It seems like the hack in
https://code.google.com/p/chromium/codesearch#chromium/tools/build/scripts/slave/recipe_modules/gclient/api.py&l=163
is the reason for this.

Example of builder successful building with the recipe (the recipe switch is currently reverted again):
http://build.chromium.org/p/client.webrtc/builders/Android%20Chromium-APK%20Builder/builds/1003
and a tester that was triggered by that build:
http://build.chromium.org/p/client.webrtc/builders/Android%20Chromium-APK%20Tests%20%28KK%20Nexus5%29/builds/1028

BUG=346198
TEST=Successful builder and tester execution using a local master and slave with device connected and https://codereview.chromium.org/240143002/ applied.

Review URL: https://codereview.chromium.org/260353002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@267123 0039d316-1c4b-4281-b951-d872f2087c98