TryChangeGerritStatus posts messages back
TryChangeGerritStatus posts messages on a Gerrit change when a build is
started or finished. The message contains a link to the build. When
build is started, an email is not sent. TryChangeGerritStatus uses REST
API to talk to Gerrit.

Tested.

TBR=vadimsh@chromium.org, iannucci@chromium.org, kmg@chromium.org
BUG=366099, 364639, 364631

TryJobGerritScheduler puts URL to the change

The change URL is assigned to the event.change.url build property.

Had to modify GerritAgent to ensure the gerrit_host does not have a
path.

R=vadimsh@chromium.org, kmg@chromium.org
BUG=366099, 364639, 364631

Review URL: https://codereview.chromium.org/262733004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@267472 0039d316-1c4b-4281-b951-d872f2087c98