WebRTC: Change webcam script to be used for standalone only.
As our Chromium bots no longer uses virtual webcams (yay!) we
can reconfigure it so it works for the WebRTC standalone builds.
It turns out that this script never managed to actually launch the
virtual webcam on Linux, since the /dev/video1 device never existed
on the standalone builders.
This was discovered after landing 
https://chromiumcodereview.appspot.com/260683002/

TBR=stip
TEST=Local testing on a bot.
BUG=webrtc:2812

Review URL: https://codereview.chromium.org/262363004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@268473 0039d316-1c4b-4281-b951-d872f2087c98