WebRTC recipe: Fix got_revision for standalone build.
The got_revision_mapping was incorrectly set to 'webrtc'
instead of 'src'. The key in the dict is supposed to point
to the path in the checkout to use for the build property to
be set (the value of the dict entry).

I realized this when deploying the first standalone builders
in https://chromiumcodereview.appspot.com/258803003/

BUG=346198

Review URL: https://codereview.chromium.org/277583005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@269051 0039d316-1c4b-4281-b951-d872f2087c98