Add builder tag to the backend compiler.
This is a hack, but I don't want to add a new compiler type or a special flag to test.py just for a few weeks.

This will enable you to add entries in the status files with
[ $compiler == dart2dart && $builder_tag == new_backend ]
or the negative version if something is only failing on the default
[ $compiler == dart2dart && $builder_tag != new_backend ]

Locally, to make use of the special tag when running you need to do something like this (i.e., to validate that tests are passing according to the new_backend status file entries).
DART_VM_OPTIONS=-DUSE_NEW_BACKEND tools/test.py -mrelease -cdart2dart --builder-tag=new_backend

I know this is sort of cumbersome, but if it is only for a few weeks it should probably be fine?

Review URL: https://codereview.chromium.org/284793004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@270065 0039d316-1c4b-4281-b951-d872f2087c98