Add oilpan category to the chromium.webkit waterfall.
Currently in chromium.webkit, the 'category' that each bot is assigned
to is set per-file in the master config, so all bots defined in that
file appear in the same category. This makes it difficult to mix and
match things across files.

This changes makes it possible to specify the category on a per-builder basis
in the master_config.Helper.Builder constructor.

Then, using that, we re-assign all of the oilpan bots into a single category
(and add the WebKit Linux Leak bot, since that is monitored by the same team).

R=iannucci@chromium.org
BUG=356849

Review URL: https://codereview.chromium.org/285093003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@270458 0039d316-1c4b-4281-b951-d872f2087c98