WebRTC: Fix Android builder slavebuilddir configurations.
When switching over the rest of the builders to recipes in
https://codereview.chromium.org/286443002/
a mistake was made for the Android builders.
Since the Android and the Android Chromium-APK type of builders
have different kind of checkouts (WebRTC standalone vs a Chromium
checkout), they cannot share the same slavebuilddir.

TBR=phoglund
BUG=346198
TEST=None

Review URL: https://codereview.chromium.org/296453008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@271628 0039d316-1c4b-4281-b951-d872f2087c98