Add a Blink GN recipe.
This adds a recipe to run GN against a Blink config of Chromium (i.e.,
Chromium + tip-of-tree Blink, the config we use on the chromium.webkit
waterfall).

This will make it feasible to properly test GN files for the Blink build
(landing and catching problems only in Blink rolls is a really bad idea).

Note that no bots use this recipe yet.

R=iannucci@chromium.org, kbr@chromium.org
BUG=376010

Review URL: https://codereview.chromium.org/291193006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@274597 0039d316-1c4b-4281-b951-d872f2087c98