Update android.api to use chrome tools via the android path
Currently the android_aosp bot symlinks the chrome source into the
android tree we want to change this to an approach where we copy the
chrome code into android tree. This means that
android_src/external/chromium_org and src will no longer
point to the same files so we have to update android.api to use the
right paths. (This can happen in advance since it they lead to the
same place currently).

BUG=373411

Review URL: https://codereview.chromium.org/323523003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@275455 0039d316-1c4b-4281-b951-d872f2087c98