Move Known Incompatibilities To Rsync Step
We need to ensure that incompatibly licenced code does not end up in the
WebView. To make the android_aosp bots check for this more similar to
the auto_merger check we want to remove all the incompatible code first
and then run build android_webview/tools/webview_licenses.py on
everything. (Before this change we rsynced some of incompatible code
into the android checkout and the used a whitelist in
known_issues.KNOWN_INCOMPATIBILITIES to ignore it.)

While we're here we also add another option to the rsync command:
--delete-excluded
Which deletes things in the android checkout which are also in
the chromium checkout but excluded by us. This stops us
accidentally using stale files which where synced from before
but we have now excluded.

BUG=311868

Review URL: https://codereview.chromium.org/326083003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@276995 0039d316-1c4b-4281-b951-d872f2087c98