Collect GIT_CURL_VERBOSE=1 traces on Win GPU Triggered bots.
The stderr of 'git fetch' will be dumped to slave/build/git_curl_trace_*.log
file. Note that because of that, error messages from 'git fetch' won't be
visible on the waterfall, and if the step fails, there's would be no clear
indication why. It's a temporary measure while User is Over Quota issue is
being investigated.

BUG=383353
R=kbr@chromium.org,iannucci@chromium.org

Review URL: https://codereview.chromium.org/347053003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@278876 0039d316-1c4b-4281-b951-d872f2087c98