V8 Buildbot: Switch arm builder to recipes.
This allows using the cross tool option for compile.py
through the chromium config.

The old way of v8 arm archiving is legacy and will be
removed. The archives are not used by anyone.

Setting the CXX and LINK env variables is not necessary as
the --crosstool option does this already in compile.py.

BUG=315202

Review URL: https://codereview.chromium.org/334263019

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@279971 0039d316-1c4b-4281-b951-d872f2087c98