Annotation: preamble is not closed by annotator
The preamble step should not be closed at the annotator level because it
is closed by LoggingBuildStep.startCommand (as a callback to
runCommand). Currently it is possible to cause the preamble step to be
closed twice, once by handleReturnCode and then by LoggingBuidStep. This does
not break anything under existing cases because the closures happen about one
after another. However, it does not work in the upcoming STEP_TRIGGER case
because the runCommand waits for async operations to finish and then calls the
callback which is BuildStep.finish, while in the meantime the log is compressed
and the second attempt to close the step fails to read the size of the
uncompressed log.

This change prevents the preamble to be closed, either by
handleReturnCode or by an annotation, such as
STEP_CLOSED/STEP_FAILURE.

R=stip@chromium.org
BUG=361841

Review URL: https://codereview.chromium.org/340843004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@280183 0039d316-1c4b-4281-b951-d872f2087c98