Added async operation support for annotated script
A step may have async operations associated with it. The step is not finished
until all its async operations complete. Accordingly, build does not complete
until all its steps complete.

This change introduces a difference between closing and finishing a
step. The step is closed by an annotation script, which means the script
will not emit anything within this step anymore, but it is up to the
master when to finish the step. It happens when all async ops complete,
if any.

From the annotation script perspective, there is no breaking change. It never
knew and never attempted to finish steps; the annotation is called
STEP_CLOSED.

The change to the runCommand makes sure that the returned Deferred fires
only after all steps finish, so the build is not completed too quickly

We need to handle the exit code after script completes, but before we start to
wait for steps to finish, so the commandComplete (called right after
runCommand completes) was replaced with scriptComplete which is called manually
inside runCommand.

Based on https://chromiumcodereview.appspot.com/340843004/
Split from https://codereview.chromium.org/277263003/

R=stip@chromium.org
BUG=361841

Review URL: https://codereview.chromium.org/346033002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@280666 0039d316-1c4b-4281-b951-d872f2087c98