Save apache access, error logs in the layout test archive on windows.
In order to help debug bug 378444, we should be archiving the
http server logs so that we can see why they're bailing out.

Historically, we didn't just archive these logs by default because
we don't need them on the bots and at least the access logs might be
large. For now, we'll just archive them on windows, and if it turns
out the logs are small, we can consider just archiving them always.

R=eseidel@chromium.org
BUG=378444

Review URL: https://codereview.chromium.org/367473002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@280699 0039d316-1c4b-4281-b951-d872f2087c98