Revert of Enable bot_update on client.webrtc (https://codereview.chromium.org/377413005/)
Reason for revert:
Webrtc Revisions aren't getting passed through, and theres no good way to fix this right now because while https://codereview.chromium.org/385513003/ looks to be the "right way" it'll tell bot_update to find an exact match on the repository.

However, since webrtc_revision refers to webrtc/trunk/, but the referenced repos are webrtc/trunk/talk and webrtc/trunk/webrtc, when you specify revision A that exists only in one of the repos but not the other, bot_update will fail loudly because the revision is non-existant in the other.

Theres no good quick fix for this issue, so I'm reverting the CL.  What we need is to combine webrtc/trunk/* into a single DEPS, and symlink the folders over.

Original issue's description:
> Enable bot_update on client.webrtc
> 
> WebRTC seems to be working okay, turning it on for this webrtc waterfall too.
> 
> BUG=261741
> TBR=agable,phoglund
> 
> Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=282187

TBR=agable@chromium.org,phoglund@chromium.org,hinoka@chromium.org
NOTREECHECKS=true
NOTRY=true
BUG=261741

Review URL: https://codereview.chromium.org/380963002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@282210 0039d316-1c4b-4281-b951-d872f2087c98