Initial checkin of the client.dart.packages waterfall.
This does not have any slaves attached yet.
This is setup to mirror the normal dart based master files looping over the release channels. This is just to get us going, we will be experimenting with different slave/repo setups, but we want to keep the somehow in sync with how we handle the normal two waterfalls.

Templates/public html is verbatim copy from client.dart, with the slaves removed from the top bars.

The ports are taken from issue 371427

Review URL: https://codereview.chromium.org/368873002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@282278 0039d316-1c4b-4281-b951-d872f2087c98