Add a status_logger for better monitoring of buildbot events.
This new status reciever is meant to simply log every single user-facing event
that occurs in a single buildbot master, and log them to disk with timestamps
where they can be grepped and analyzed. This is motivated by the desire to see
if particularly large or particuarly many simultaneous logs were the cause of
the recent tryserver high load and failures.

R=cmp@chromium.org, dnj@chromium.org

Review URL: https://codereview.chromium.org/397173005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@284260 0039d316-1c4b-4281-b951-d872f2087c98