Make the Win-Clang Dbg FYI builder do a shared_library build
This is the default for Debug builds. We were previously forcing it
to do a static build because Clang didn't have good enough dllimport/export
support, but that should work now.

The shared library build doesn't work with NaCl yet, so disable it for now.

BUG=82385

Review URL: https://codereview.chromium.org/407003003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@284538 0039d316-1c4b-4281-b951-d872f2087c98