Remove 'changelog.xml' from archive
Remove the generation of 'changelog.xml' from the uploaded
archive in 'chromium/archive_build'. It isn't implemented for
'git' repositories and doing so is nontrivial. A preliminary search didn't
reveal any infrastructure that actually uses it.

This change will remove the files; feel free to revert if it turns out that they're actually in use.

BUG=396140
TEST=None
R=agable@chromium.org, iannucci@chromium.org, maurel@chromium.org

Review URL: https://codereview.chromium.org/409113004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@284749 0039d316-1c4b-4281-b951-d872f2087c98