Fail try runs on with-patch failure in telemetry.
This restores the behavior from before r284680. That cl changed how telemetry tests run so that they're started after triggering swarming. A side effect of that is they're rerun without the patch on failure. However until we have JSON output, we have a hardcoded string of the failing tests. That means that a cl could introduce new failures in telemetry and it would land.

BUG=359521

Review URL: https://codereview.chromium.org/414553002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@284772 0039d316-1c4b-4281-b951-d872f2087c98