Add blink recipe for component builds.
This recipe can be used by components like v8 to verify
blink tests with a low false positive rate. Similar to a
trybot, this recipe compares test failures from a build with 
a current component revision with test failures from a build 
with a pinned component revision.

The contents of this recipe are mostly copied from the
blink_trybot recipe and stripped a little as not everything
is needed here.

BUG=376950

Review URL: https://codereview.chromium.org/419803002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@285879 0039d316-1c4b-4281-b951-d872f2087c98