V8 Buildbot: Repair shown revision on blink downstream bots.
The console view should always use the revision of the first
sync attempt. In case of test failures, the recipe syncs to
the pinned revision for comparison, but it should not be used
on the waterfall.

This also adds me to the OWNERS which I forgot in the
initial CLs.

BUG=376950

Review URL: https://codereview.chromium.org/426153005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@286488 0039d316-1c4b-4281-b951-d872f2087c98