Enable bot_update for chromium codesearch
This changes the codesearch builder on chromium.fyi to split out all of its
dependencies into top level solutions (instead of one massive build/ dep) and
turns on bot_update.

The reason this is done in conjunction with turning on bot_update is because
gclient is not able to have slashes in the solution name.

BUG=393781,351896

Review URL: https://codereview.chromium.org/420603015

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@286555 0039d316-1c4b-4281-b951-d872f2087c98