Fix v8 got_revisions
So the way got_revisions work right now is a bit scary.  Instead of using the
normal got_revision_mapping codepath, at the end of the GClient step, chromium_steps.py's
Gclient step catches the got_revision emission, tosses it on the ground, and replaces
it with got_v8_revision.

The problem with this is that it completely bypasses the existing and sanctioned
"got_revision_mapping" passed into the step, and hacks it at the end of a step
on the master side to mask what gclient has already emitted to something else.
But this only happens on the gclient step, and is not portable (and we don't want
it to be portable) over to bot_update.  This is why bot_update client.v8 builders
are currently emitting a chromium got_revision instead of a V8 got_revision.

This CL adds an option to do got_revision_mapping_overrides at the ChromiumFactory
initilization level so that bot_update can also get the same revision_mapping.

BUG=393784

Review URL: https://codereview.chromium.org/426243009

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@286655 0039d316-1c4b-4281-b951-d872f2087c98