Reland support for retrying of telemetry_unittests again (take 3).
The previous two commits that got reverted - r286260 and r285595 -
were actually both fine, but they revealed bugs in chrome code
that caused badness.

This patch is the same as the others, with one additional check to
ensure that the json results files are consistent with the
exit code from the python command.

R=iannucci@chromium.org
BUG=359521

Review URL: https://codereview.chromium.org/433703002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@286821 0039d316-1c4b-4281-b951-d872f2087c98