Add a GitilesStatus HtmlResource object
Added a base.HtmlResource object that provides the status of a
GitilesPoller.

R=szager@chromium.org

Add support for master custom endpoints

Added support for a master to pass through a dictionary of custom
endpoints.  The dictionary uses the URL string as a key and the
base.HtmlResource object as a value.

R=agable@chromium.org, iannucci@chromium.org
BUG=

Review URL: https://codereview.chromium.org/437163004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@287368 0039d316-1c4b-4281-b951-d872f2087c98