Revert of Make ninja display elapsed compile time by default. (https://codereview.chromium.org/268833006/)
Reason for revert:
This made the build pages more noisy and happened to not be as useful as hoped in practice.

If someone wants to use this to debug something, please add it temporarily while you debug, and then remove it again when you're done.

Original issue's description:
> Make ninja display elapsed compile time by default.
> 
> This will help debug slow or hanging compiles.
> 
> BUG=none
> 
> Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=269042

Review URL: https://codereview.chromium.org/405733002/


git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@287547 0039d316-1c4b-4281-b951-d872f2087c98