client.webrtc: GitilesPoller
Replaced SVNPoller with GitilesPoller in client.webrtc, so only
trunk/webrtc changes are polled because that's what we check out. Before this
change revisions of the upper directory were polled too, which is more than we
need to build and bot_update breaks on this because it cannot find with the
exact svn rev among git commits in the git trunk/webrtc mirror.

This also makes the recipe use HEAD for both talk and webrtc, since it's
difficult to know what the right polling behavior SHOULD be based just on the
'revision' property.

Carried over from https://codereview.chromium.org/455103002/

R=kjellander@chromium.org, iannucci@chromium.org
BUG=393786

Review URL: https://codereview.chromium.org/454053003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@288827 0039d316-1c4b-4281-b951-d872f2087c98