Added build ID support, condition Property class.
Added support to ChromiumOS master to optionally pull the 'build_id:'
configuration line out of a manifest 'git' commit and supply it to 'cbuildbot'
as the '--master-build-id' parameter if present.

To support this, added a 'ConditionalProperty' class to Chromium 'master_utils'
that allows greater freedom for property-based command-line generation. This
also addresses one of the TODOs in the ChromiumOS factory.

BUG=398267
TEST=localtest

Review URL: https://codereview.chromium.org/453513002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@288960 0039d316-1c4b-4281-b951-d872f2087c98