Skia recipe: Add compile() to default_flavor
This is the first CL in a series which will add
compile support for Android, ChromeOS, and NaCl.

For background, Skia does the same basic set of
build steps in slightly different ways depending on the platform:
- compile code
- run test A
- run test B
- etc

We use the flavor utils classes to determine *how* to run those
steps.  For example, the default flavor is for running code on normal,
desktop machines. Therefore, we just compile normally using ninja and
then run the tests from the out/ directory.  For Android, we compile
using an Android toolchain, and run the tests using a script which pushes
the executable to the device and runs over ADB.  The flavor abstraction
allows us to give a high-level "compile target A" command and have the
flavor utils determine how that needs to be done for the given platform.

BUG=skia:761

Review URL: https://codereview.chromium.org/445363004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@288977 0039d316-1c4b-4281-b951-d872f2087c98