WebRTC: Fix Android APK testers in client.webrtc
This CL intends to solve the currently failing
Android Chromium-APK testers in client.webrtc.

In https://codereview.chromium.org/454053003/
iannucci@ changed polling to the Git mirror to make
bot_update work properly for the Chromium-APK bots.
The testers also need to sync HEAD similar to the builders
to avoid running into the problem that some revisions in the
WebRTC and Libjingle mirrors do not exist in both mirrors.

The revision used for package and extract build must still be
the one passed from the poller. Otherwise there will be failures
when new revisions are checked in between the poller has triggered
the builders and the actual bots perform their sync steps.

BUG=393786
TBR=iannucci@chromium.org

Review URL: https://codereview.chromium.org/468823002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@289273 0039d316-1c4b-4281-b951-d872f2087c98