Re-land r289084 - "Make telemetry*_unittests upload to test-results.appspot.com"
This version differs from the prior attempt by

a) changing the logic so that we only upload the results if we're not running on a trysever.

b) retrieving the repo revisions from the data that bot_update got, rather than whatever was passed in as part of build_properties. Turns out that the build_properties would've only had 'parent_got_revision' set and not 'got_revision', so we were likely referencing keys that didn't exist.

As part of (b), this changes adds a "properties" property to the bot_update recipe module so that we can retrieve what bot_update actually did.

R=agable@chromium.org
BUG=323212, 403564

Review URL: https://codereview.chromium.org/476503002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@289434 0039d316-1c4b-4281-b951-d872f2087c98