Add more builders to client.skia
The point is to test various types of bots to make sure they're working as expected:

Build-Ubuntu13.10-*: Ensure that we build correctly for Android, ChromeOS, and NaCl.
Build-Mac10.7-*: Ensure that bots in Skia Lab can connect to this master.
Test-Android-*: Ensure that the Android flavor works and continues to work.

BUG=skia:761

Review URL: https://codereview.chromium.org/472983002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@289665 0039d316-1c4b-4281-b951-d872f2087c98