https://codereview.chromium.org/469533002 made a change where which caused all gtest_filter-specifying invocations to always run with one job, which caused problems for the WebRTC tests (and seemed to trigger a bug where tests sometimes hang if they're first to execute).
To work around this, I explicitly specify the number of jobs to use
in the WebRTC bot config. Specifying bot mode will not help since
https://codereview.chromium.org/469533002 only looks at gtest_filter
and not the bot mode flag. An alternative solution is to say run with
one job if gtest filter is specified and bot mode is NOT specified.

Depends on https://codereview.chromium.org/485733002/

BUG=402089

Review URL: https://codereview.chromium.org/480053002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@290599 0039d316-1c4b-4281-b951-d872f2087c98