WebRTC: Add Mac as deps OS for iOS recipe.
In https://codereview.chromium.org/477803003/ the recipe got
target_os = ['ios'] added in order to pull the iOS deps. This is
not enough since WebRTC's iOS build depends on src/third_party/openmax_dl,
which is not synced for iOS in Chromium. Solve this by adding 'mac' as well
to the target_os list.

BUG=webrtc:2863
TEST=Verified that openmax_dl is not synced when target_os = ['ios'] and
that it is synced (and symlinked) when target_os = ['mac', 'ios']

Review URL: https://codereview.chromium.org/481833004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291093 0039d316-1c4b-4281-b951-d872f2087c98