Switch chromiumos.chromium to use GitilesPoller in Git mode.
The SyncChrome step is failing because it tries to access /+/<svn-rev>/ urls
on googlesource. That obviously doesn't work. It's getting the svn-rev from
the revision build property, which is set by the poller. This switches it to
start using git now, so that the revision property is a git hash instead.

TBR=dnj@chromium.org, szager@chromium.org
BUG=406454

Review URL: https://codereview.chromium.org/499563005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291457 0039d316-1c4b-4281-b951-d872f2087c98