Revert "In buildspec mode, ignore got_revision and just emit the primary repo as the revision"
This reverts commit c98cf411edb8a9cf7f7b594a4032865cec9755da.

Not necessarily a problem for the build, but this is causing buildbot to
display weird reivision information, which is confusing and might cause
people to think something is wrong when it's not, or to go on wild goose
chases when investing real problems. Reverting while investing a
solution without those side-effects.

TBR=amineer@google.com
BUG=406955

Review URL: https://codereview.chromium.org/498243003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291607 0039d316-1c4b-4281-b951-d872f2087c98