Infra recipe how to run presubmit
Once this is done, we can make the presbumit
not actually run tests.py, so that we don't run
it twice.

Originally landed as:
https://codereview.chromium.org/483023003
But then reverted as:
https://codereview.chromium.org/494733003/

The underlying issue of the patch not being applied was
caused by the infra tryjob being run on the
chromium.linux master and defaulting to
root='src' causing all patches to fail to apply.

That issue is in the process of being fixed.

BUG=402776
TBR=eseidel@chromium.org

Review URL: https://codereview.chromium.org/506573004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291622 0039d316-1c4b-4281-b951-d872f2087c98