Removes chrome_run from compile_targets
Turns out depending upon chrome_run is a bad idea. We don't always
generate it. So, instead I'm passing any targets that conditional
tests depend upon to the filter step as an exe to look for. This way,
if there is a match we're guaranteed to get the target output
matching_exes. This allows for removal of the chrome_run target.

BUG=407419
TEST=none
R=iannucci@chromium.org

Review URL: https://codereview.chromium.org/509643002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291657 0039d316-1c4b-4281-b951-d872f2087c98