Move 10 builders from rel to dbg windows.
We are running into queues of jobs on the debug builder. We have a lot of spare capacity on the windows release builders with the switch to swarming and improvements in only running the tests which we need. As a result, the previous fractional split doesn't hold anymore.

BUG=407792

Review URL: https://codereview.chromium.org/509663002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291662 0039d316-1c4b-4281-b951-d872f2087c98