Add an OWNERS file for blink_roller
It's silly that Eric Boren and I wrote all this
code and yet can't review it.

The root OWNERS file appears to be a list of all of
the chrome infrastructure team.  I don't know the
rest of the code in build/ so just creating a new
OWNERS file for this directory.

I'm also happy to move this script to infra/
instead, but I'd rather move the whole of the
build/ repo to infra/ instead of moving one script
at a time.

BUG=406867

Review URL: https://codereview.chromium.org/515183003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291708 0039d316-1c4b-4281-b951-d872f2087c98