Remove endure waterfall and special endure-processing code.
This CL is intended to remove everything in the build repo related to endure.

The planned next incarnation of endure will be a simpler and shorter-duration
benchmark which should be run on the chromium.perf waterfall.

As of the first patch, there are still some occurrences of the string "endure"
in some recipe .expected json files, e.g.
scripts/slave/recipes/android/perf.expected.
Not sure if these should be removed too.

BUG=397206

Review URL: https://codereview.chromium.org/499143002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291709 0039d316-1c4b-4281-b951-d872f2087c98