WebRTC: Remove Chrome OS bot.
This bot is not a real Chrome OS but only a Linux
bot that builds and tests with the GYP_DEFINES flag
chromeos set to 1. I have verified that we don't have
a single condition in the WebRTC GYP files using this
variable, so I don't think this bot serves any purpose.

BUG=
TBR=phoglund@chromium.org

Review URL: https://codereview.chromium.org/522653003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291723 0039d316-1c4b-4281-b951-d872f2087c98