V8 Buildbot: Add simple v8 releases recipe.
This is a wrapper recipe that bootstraps the necessary
checkouts and calls the v8-side script.

The v8 checkout needs to be a git svn clone that is
bootstrapped from a zip file.

The v8 side releases.py script collects information about
chromium and v8 revisions and branches for external
presentation. The script does not write to any repo.

BUG=408523
TBR=hinoka@chromium.org

Review URL: https://codereview.chromium.org/515553002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291755 0039d316-1c4b-4281-b951-d872f2087c98