Added 'etc' builder to ChromiumOS Tryserver.
Added an internal and external 'etc' builder to the ChromiumOS tryserver. The
builder is scheduled through a special try job command that allows the requester
to specify explicit build names, even if a build-specific bot isn't associated
with it. This will allow developers to test builds before they are added to
official waterfalls.

BUG=348660
TEST=local
  - Scheduled internal 'etc' build through manual Git commit; observed correct
    'cbuildbot' invocation.

Review URL: https://codereview.chromium.org/559193002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291944 0039d316-1c4b-4281-b951-d872f2087c98