WebRTC: Pin SDK version to 10.9 for Mac and iOS.
Recent changes in WebRTC have introduced dependencies on
the AVFoundation library, which is not accessible in the
default SDK (which is still 10.6). Setting SDK to 10.7 would
be enough to fix this particular issue, but our engineers
would rather want to pin to 10.9 to ensure we catch problems
earlier on (see referenced bug for discussion).

BUG=webrtc:3120
TESTED=Local builds setting mac_sdk=10.9 for Mac and iOS.

Review URL: https://codereview.chromium.org/573673002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@291964 0039d316-1c4b-4281-b951-d872f2087c98