Added 'etc' builder to ChromiumOS Tryserver.
Added an internal and external 'etc' builder to the ChromiumOS tryserver. The
builder is scheduled through a special try job command that allows the requester
to specify explicit build names, even if a build-specific bot isn't associated
with it. This will allow developers to test builds before they are added to
official waterfalls.

This is the second version of the CL; it differs from the first in that it
constrains the namespace of the builds that are allowed to be invoked.

BUG=348660
TEST=None

Review URL: https://codereview.chromium.org/571523003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292007 0039d316-1c4b-4281-b951-d872f2087c98