Fix compile optimization in analyze.py in Android trybots.
It looks like when the initial set of compile targets is an empty list, which is what happens for android_dbg_tests_recipe, then the intersection of it and the supplied list from analyze.py is empty. So we end up building everything. Fix this by detecting when there's no list initially, and just building whatever analyze.py says.

BUG=405584, 109173

Review URL: https://codereview.chromium.org/586803002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292044 0039d316-1c4b-4281-b951-d872f2087c98