Avoid reusing the pgd file from a previous PGO build.
If we don't do this then the pgd file become bigger and bigger and the build time increase a lot.

In theory it'd be interesting to reuse the previous pgd file, as it might give more hint to the linker and what should be optimize. In practice it's pretty useless because we always run the small set of tests.

R=iannucci@chromium.org

BUG=309849

Review URL: https://codereview.chromium.org/601373002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292159 0039d316-1c4b-4281-b951-d872f2087c98