10% experiment on GOMA_SEND_SUBPROGRAM_SPEC.
In the previous config, we have tried adding GOMA_SEND_SUBPROGRAM_SPEC
for selected hosts on Linux, Mac, and Win.
Now, no subprogram mismatch is observed.

To move this forward, let me apply this to "randomly" selected host.
To make it easy to trace the broken host on this experiment, the selection
is done with the value calculated from its hostname.

Review URL: https://codereview.chromium.org/608693002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292169 0039d316-1c4b-4281-b951-d872f2087c98