Move android_webview_aosp_perf to chromium_android
Switch from using the WebView custom stuff in
recipe_modules/android to using what everybody else uses
recipe_modules/chromium_android.

This fixes the problem which kept killing the bot (We "used goma"
when we compiled but not when we gyped which caused us to try to do a
-j50 build locally) and has the nice side effect of removing a whole
lot of code.

BUG=411843, 391488

Review URL: https://codereview.chromium.org/611073002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292219 0039d316-1c4b-4281-b951-d872f2087c98