Revert "Make timeouts show up as purple build status on buildbot"
This reverts commit bc604d48af713c9e43b6641878be5b00869876cf.

Reason for revert:
This does what we want, but not cleanly. I've seen multiple cases of builds that are purple but there's no indication anywhere that the reason for them being purple is a timeout. Furthermore, the step that timed out is still in progress (yellow). This is confusing sheriffs.

BUG=419038
TBR=stip,phajdan

Review URL: https://codereview.chromium.org/620583002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292220 0039d316-1c4b-4281-b951-d872f2087c98