Remove the [swarming] prefix on collect steps.
From the user standpoint, the test ran. It doesn't matter this much if the test
was run locally or on Swarming.

The [trigger] prefix is kept on triggering, along with relevant links to the
Swarming task(s) so all the data is there to diagnose in case of swarming
internal failure.

R=vadimsh@chromium.org
BUG=418844

Review URL: https://codereview.chromium.org/606083005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292222 0039d316-1c4b-4281-b951-d872f2087c98