Move cleanup from the end of the script to the start.
Currently the cleanup only happens if the script runs successfully.
This means for example that when the master restarts and the script is aborted,
that the huge source archive will not be cleaned up locally, eventually leading
to the buildbot having no disk space left.
This CL moves the cleanup to the start of the script, thereby making sure that
all data from previous runs of the script are cleaned up properly before
creating new data.

BUG=419340

Review URL: https://codereview.chromium.org/621533003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292242 0039d316-1c4b-4281-b951-d872f2087c98