Remove win8_aura bot in favor of win8_chromium_{rel,dbg} bots.
The win8_chromium_* bots are closer to the main waterfal configuration, are on
recipes, have all of the spiffy modern patch application and test filtering logic.
Additionally, the win8_aura is not in the default trybot set, and steals resources
from the win8_chromium* bots when someone triggers it manually.

R=cpu@chromium.org, phajdan.jr@chromium.org, scottmg@chromium.org, wfh@chromium.org
BUG=373973

Review URL: https://codereview.chromium.org/620883003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292261 0039d316-1c4b-4281-b951-d872f2087c98