Re-land https://codereview.chromium.org/577643004, updated to work with recipes.
This CL parses the arguments passed to runtest.py, and assumes if the command is "run_benchmark" then it is telemetry under recipes. It parses the run_benchmark command line, changes the output-format from buildbot to chartjson and adds a --output argument. Then it does the same set of changes as https://codereview.chromium.org/577643004 to modify the chartjson output and send it to the perf dashboard.

When this has landed and we are unlikely to revert, we can submit https://codereview.chromium.org/595793005/ and then remove the code here that switches the --output-format argument from buildbot to chartjson. Modifying the argument in this CL means we will only have to revert/reland one CL if this causes any problems.

BUG=420044
NOTRY=true

Review URL: https://codereview.chromium.org/623313002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292343 0039d316-1c4b-4281-b951-d872f2087c98