Fix the gtest_filter used for browser_tests on ARM: remove excess quotes
If the string '--gtest_filter="*NaCl*"' were passed to the shell, the
shell would remove the ""s.  However, I don't think this string is
being passed to the shell, so gtest is looking for test names
containing ""s, which causes it to run zero test cases inside
browser_tests on the ARM trybot.

BUG=420758
TEST=Speculative change -- not tested

Review URL: https://codereview.chromium.org/630213002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292358 0039d316-1c4b-4281-b951-d872f2087c98