Mostly revert 2d50c2cd "Added analyze support to GPU recipe, thanks mainly to @zmo."
On the real tryserver waterfall, analyze was repeatedly saying that no
tests had to run. Need to figure out the discrepancy between the real
script and the simulated test data. Take out the code until this is
better understood. A few of the code changes remain compared to the
previous state.

BUG=421954

Review URL: https://codereview.chromium.org/640253005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292389 0039d316-1c4b-4281-b951-d872f2087c98