Use Skia favicon.ico in master.client.skia.
The Skia favicon.ico was created with 292388 "Add Skia favicon" but
this did not actually cause the icon to be used. The browser is
still looking for the favicon at the server root, and the
client.skia master is running at /p/client.skia.

This change adds a 'link' in the 'head' of client.skia pages which
points to the correct favicon.

R=borenet@google.com

Review URL: https://codereview.chromium.org/651463002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292436 0039d316-1c4b-4281-b951-d872f2087c98