Master emits 'CHROME_HEADLESS' for 'git-number'
A change was made to 'git-number' to stop developers from accidentally relying
on it. This change requires the 'CHROME_HEADLESS' environment variable, a
varaible which is defined on slaves, to be present in order to run.

Have the master also emit this environment variable when calling 'git-number'.

BUG=chromium:423038
TEST=None
R=iannucci@chromium.org, tandreii@chromium.org

Review URL: https://codereview.chromium.org/664553003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292495 0039d316-1c4b-4281-b951-d872f2087c98