No longer retry builds with a RETRY status.
This is a pretty fundamental change to the way that Builder works. Previously,
if the slave was lost, the build master would be responsible for retrying the job.

However, nowadays, we either:
  * want to try the next polled job (main waterfalls)
  * want to have CQ try some almost-but-not-quite-the-same job

In no event that we can think of is it actually advantageous to have the
buildmaster do this retry. Also unfortunately (?) the Builder object is directly
instantiated by the Master, (e.g. there's no BuilderFactory!), so modifying
Builder directly was the quickest/best way to deal with it.

R=cmp@chromium.org, dnj@chromium.org, kbr@chromium.org, sergeyberezin@chromium.org
BUG=424226

Review URL: https://codereview.chromium.org/664583005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292500 0039d316-1c4b-4281-b951-d872f2087c98