Only RETRY on disconnections when the master is restarting.
This makes it so that the RETRY status will only be set when picking up
disconnections /when the master is restarting/.

This way regular slave disconnections (e.g. the slave panic'd) count as
EXCEPTIONS, but restarting the master will still RETRY all dropped jobs.

R=cmp@chromium.org, dnj@chromium.org, kbr@chromium.org, sergeyberezin@chromium.org
BUG=424226

Review URL: https://codereview.chromium.org/652843003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292502 0039d316-1c4b-4281-b951-d872f2087c98