Add android and chromeos builders to client.mojo
This adds android and chromeos builders to the client.mojo waterfall,
configured to run on the same vm as mojo linux. The average cycle time
of the linux bot is under 90 seconds and android/chromeos will be build
only to start so they should fit fine.

R=dpranke@chromium.org

Review URL: https://codereview.chromium.org/676283002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292709 0039d316-1c4b-4281-b951-d872f2087c98