Add '-v' to actual-results.json gs upload.
This fixes an asymmetry between the actual result .png files (which
are versioned by a hash of pixels in name) and the actual-results.json
file. This file actually is being versioned in gs, but we do not
indicate in the build output which one we uploaded. This should fix
that.

The primary reason for this is that, with this extra information, it
should be possible to point rebaseline_server at a reitveld patch and
have it scrape the actual-results.json produced by tryjobs and use
that as the basis for rebaselining.

Review URL: https://codereview.chromium.org/699653002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292804 0039d316-1c4b-4281-b951-d872f2087c98