Add a --recursive flag to compare_build_artifacts and always look at the content of the *_apk subdirs.
The new flag makes the debugging much faster when the .isolated refers to some files that are not in the product directory.

On Android the apks are not in the product directory, they're in a directory called $(test_name)_apk/ (e.g. net_unittests_apk/net_unittests.apk)

BUG=314403

Review URL: https://codereview.chromium.org/702083004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292850 0039d316-1c4b-4281-b951-d872f2087c98