Flip the Windows Chromium GN bots to by x64 by default.
And re-enable the compile on the tryservers (but not the waterfall
bots).

It looks like it will be easier to get the all-64-bit build working
at first, so we'll do that and then go back to making the 32-bit build
work as well.

TBR=scottmg@chromium.org
BUG=432375

Review URL: https://codereview.chromium.org/730203004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292980 0039d316-1c4b-4281-b951-d872f2087c98