Display telemetry chartjson output and result data for android on waterfall.
This enables chartjson output from telemetry tests and displays the perf results on the waterfall, along with a link to the json.output that was sent to the dashboard.

The changes to android's test_runner need to land first, to teach it to save and output the chartjson results.

Needs to land first: https://codereview.chromium.org/759753008/

BUG=422174

Review URL: https://codereview.chromium.org/753073007

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293167 0039d316-1c4b-4281-b951-d872f2087c98