Update DynamoRIO and Dr. Memory package steps to git.
Previously we relied on got_revision matching the package revision, but
that is no longer the case with git.  We instead use the short git hash as
the build number (for the Windows build this gets truncated in some cases
but otherwise works).  We also rename both sets of packages to more closely
match the official package naming scheme.

To satisfy the mock build tests we have to initialize the properties we're
setting in build steps up front in the masters as the mock build does not
run the steps.

TBR=zhaoqin@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/778653003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293235 0039d316-1c4b-4281-b951-d872f2087c98