Run html_viewer_unittests on the Linux GN bots.
There are a class of mojo services that live in the Chromium repo
that have unittests. We need to run these tests on bots somewhere,
but these tests are only built in GN builds, so we can't add these
tests to the regular bots.

There is work in progress to set up new bots for the mojo services,
but for the meantime we'll run them on the GN bots.

R=jamesr@chromium.org, aa@chromium.org
BUG=493215

Review URL: https://codereview.chromium.org/773423003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293258 0039d316-1c4b-4281-b951-d872f2087c98