Fix swarming dimension not being set for _ng bots.
sync_and_configure_build does this for the main waterfall, but for trybots it's only called for the main builder.

As part of this, since swarming dimensions are global, I changed the _ng bot lists to only have one tester. All of them are currently like that, and it makes it clearer (i.e. otherwise it would be ambiguous which swarming dimension got used).

BUG=415647

Review URL: https://codereview.chromium.org/785663004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293296 0039d316-1c4b-4281-b951-d872f2087c98