Add a new buildbot-tool generator for master configs.
We eventually want to get to a point where we have a
single stock config for buildbot masters (and slaves,
and recipes) where as much as possible is configured
from a single data file plus client-side recipes. This
would allow us to change master configs by simply 
updating scripts and either noting that the data file
changed (and having buildbot reconfigure itself
automatically), or by typing 'make reconfig' to force
it to do so without aborting any existing state.

This tool is perhaps the first step on that path: it 
uses a single (json) data file and a set of templates,
and adds a generator utility called 'buildbot-tool'
that can generate a static, checked-in master config.

None of the 'make reconfig' stuff works yet, and we
only support one very simple kind of waterfall
that uses a single Gitiles poller.

R=iannucci@chromium.org, luqui@chromium.org
BUG=356974

Review URL: https://codereview.chromium.org/761253003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293339 0039d316-1c4b-4281-b951-d872f2087c98