Split the chromium_gn.py in two for chromium.mojo.
We will be setting up a new chromium.mojo master and new bots
dedicated to running tests for the Mojo services that live in
the Chromium repo. We do not want to re-use the normal GN bots,
in order to minimize confusion.

This CL creates a new 'chromium_mojo.py' recipe and refactors much
of the code shared by the chromium_mojo and chromium_gn recipes
into routines that move into the chromium recipe_module.

R=jamesr@chromium.org, iannucci@chromium.org
BUG=439281

Review URL: https://codereview.chromium.org/791733005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293388 0039d316-1c4b-4281-b951-d872f2087c98