Make slaves.cfg optional for buildbot masters.
We are (hopefully) moving to a world where buildbot masters are
configured from a single config file that contains no executable
code. We call this new file 'builders.pyl', and you can generate
a new master using it and the new 'buildbot-tool' in scripts/tools;
it relies on code in scripts/master/master_gen.py to actually
configure the master.

The 'slaves.cfg' files are a strict subset of the information contained
in a 'builders.pyl' file and, accordingly, there is no real need for
both files to be checked in.

This change updates the code base to move all non-master-specific
references to 'slaves.cfg' into scripts/chromium_utils.py and to
make sure that we read the data from builders.pyl correctly and
do not rely on the existence of slaves.cfg.

Individual master configs that still use slaves.cfg may continue
to do so in very tightly constrained ways.

R=iannucci@chromium.org, agable@chromium.org, luqui@chromium.org
BUG=356974

Review URL: https://codereview.chromium.org/801963003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293409 0039d316-1c4b-4281-b951-d872f2087c98