Add a new chromium.mojo master.
This adds a new master to master1. It will run as one of the
main chromium waterfalls, and will build the Mojo services that
live in the chromium repo and run the tests for them.

These services can only be built using GN, and hence we cannot
simply build and test them on the normal builders.

The master config is generated from builders.py by the new buildbot-tool
script.


R=jamesr@chromium.org, aa@chromium.org, luqui@chromium.org
BUG=439281

Review URL: https://codereview.chromium.org/782043003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293411 0039d316-1c4b-4281-b951-d872f2087c98