Fixes for the chromium_gn_upload recipe, take 2.
1) The linux_chromium_gn_upload_x64 bot was incorrectly declared
as a 32-bit bot.

2) We were not explicitly setting the "cpu_arch" GN argument on
Linux or Mac, and we should've been.

These issues, in conjunction with the buggy default handling of cpu_arch
on Linux (that was recently fixed but not yet rolled in), were causing
the linux bots to get the wrong sysroots and thus causing compile
failures.

TBR=scottmg@chromium.org
BUG=430215

Review URL: https://codereview.chromium.org/816873002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293453 0039d316-1c4b-4281-b951-d872f2087c98