Fix bugs in builders.pyl handling discovered w/ chromium.mojo.
chromium.mojo is the first master created with buildbot-tool
and the new 'builders.pyl' config file syntax.

In starting it up, we discovered that the 'buildbot' file in
the master directory is actually needed (we run it when creating
and upgrading databases), and that there was a bug in figuring
out the master to connect to from a slave that was hidden when one
uses TESTING_MASTER to test things locally.

In addition this patch removes a bunch of dangerously-duplicated
code between chromium_utils.py and buildbot_tool.py that was
responsible for actually parsing the file.

R=iannucci@chromium.org, johnw@chromium.org, luqui@chromium.org
BUG=356974, 439281

Review URL: https://codereview.chromium.org/814333002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293479 0039d316-1c4b-4281-b951-d872f2087c98