Move oilpan perf bots to chromium.perf.fyi and restructure chromium.perf.fyi
The oilpan perf bots are flaky and causing a lot of headaches for perf sheriffs.
We decided to move the bots to chromium.perf.fyi to reduce the noise.

We also intend for chromium.perf.fyi to be an experimental staging location for any new
benchmark on its way to chromium.perf. Therefor, we decided to restructure it to look
like chromium.perf

BUG=433355

Review URL: https://codereview.chromium.org/807013003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293493 0039d316-1c4b-4281-b951-d872f2087c98