Set project on repo_poller changes.
Remove the ability to set the project attribute on RepoPoller
objects.  Instead set the project as detected by the git log so we
can use it for Gerrit links or ChangeFilters.  It seems that category
and the current implementation of the project attribute are more or
less redundant.  RepoPoller has two different definitions of project
and being able to identify the repo project seems more useful than an
attribute set when RepoPoller is instantiated.

There are only two other users of RepoPoller and neither of them
set the project.

R=nodir@chromium.org, agable, ngeoffray, szager
BUG=446650

Review URL: https://codereview.chromium.org/808753002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293535 0039d316-1c4b-4281-b951-d872f2087c98