Buildbucket: handle RETRY and SKIPPED builds
IF a build status is RETRY, do not do anything with it. This status is
used when master is restarted. The build should not be marked as failed.

If SKIPPED, do not do anything either: the build lease will expire on
its own. This does not happen in practice.

R=sergeyberezin@chromium.org, sheyang@chromium.org, vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/788773008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293657 0039d316-1c4b-4281-b951-d872f2087c98