buildbucket: track leases instead of slaves
buildbot-buildbucket used to check if there are available slaves before
leasing builds. The problem is that if there are other schedulers that actively
put build requests, the chances are buildbucket will never schedule
a build because there will be a lot of pending build requests and/or all
build slaves will be busy.

The solution is to track the number of currently held build leases. Now
buildbot peeks builds until it reaches the maximum number of build leases,
which is the number of build slaves by default.

Also this change includes fix to heartbeats: buildbot's buildETAUpdate
notification is not called if ETA is unknown, so now we are not relying
on that anymore, but just sending heartbeats for all currently held
leases every minute.

Also updated README

R=sheyang@chromium.org, sergeyberezin@chromium.org, vadimsh@chromium.org
BUG=449192

Review URL: https://codereview.chromium.org/799933009

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293714 0039d316-1c4b-4281-b951-d872f2087c98